added Step 4 screen
