+++ src/locale/translations/index.js
@@ -1,1 +1,1 @@
-export default JSON.parse("{\"en\":{\"Home\":{\"AppName\":\"COVID Alert\",\"UserStopped\":{\"Title\":\"You turned off the app\",\"CTA\":\"Turn back on\"},\"NoExposureDetected\":{\"AllSetTitle\":\"You’re all set\",\"RegionCovered\":{\"Title\":\"No exposure detected\",\"Body\":\"You have not been near anyone who reported a COVID-19 diagnosis through this app.\",\"GuidanceUrl\":\"https://www.canada.ca/en/public-health/services/publications/diseases-conditions/covid-19-how-to-isolate-at-home.html\",\"AllSetBody\":\"Your phone is now running Bluetooth to collect codes from nearby phones.\"},\"NoRegion\":{\"Title\":\"No exposure detected\",\"Body\":\"You have not been near anyone who reported a COVID-19 diagnosis through this app.\\n\\nHowever, in some provinces and territories, people cannot yet report COVID-19 diagnoses through this app.\",\"AllSetBody\":\"We’ll let you know as soon as people outside Ontario can report a diagnosis through COVID Alert.\\n\\nIt’s still helpful to keep COVID Alert on.\\n\\nThat way, when people are able to report a diagnosis, you’ll find out if you were near them.\"},\"RegionNotCovered\":{\"Title\":\"No reporting yet in your area\",\"Body\":\"People in your province or territory are not yet able to report a COVID-19 diagnosis through this app.\\n\\nIt’s still helpful to turn COVID Alert on. That way, when people are able to report a diagnosis, you’ll be notified if you were near them.\",\"GuidanceUrl\":\"https://www.canada.ca/en/public-health/services/publications/diseases-conditions/covid-19-how-to-isolate-at-home.html\",\"AllSetBody\":\"We’ll let you know as soon as people in your area can report a diagnosis through COVID Alert.\\n\\nIt’s still helpful to keep COVID Alert on.\\n\\nThat way, when people are able to report a diagnosis, you’ll find out if you were near them.\"}},\"ExposureDetected\":{\"Title\":\"You’ve been exposed\",\"Body1\":\"You’ve been near at least one person who reported a COVID-19 diagnosis through the app. You were near them for at least 15 minutes.\",\"Title2\":\"What next?\",\"Body2\":\"You’re at risk of being infected.\",\"Notification\":{\"Received\":\"Notification received\"},\"NegativeTest\":{\"CTA\":\"Negative test?\"},\"Dismiss\":{\"Title\":\"Negative test result?\",\"Body\":\"Dismiss this alert to receive any new exposure notifications. Dismissing this alert will not delete the exposure logs on your phone.\",\"Body2\":\"You can also dismiss the alert if you’re following your area’s guidance.\",\"CTA\":\"Dismiss alert\",\"Confirm\":{\"Body\":\"Are you sure you want to dismiss this alert?\",\"Accept\":\"Yes, dismiss\",\"Cancel\":\"Cancel\"}}},\"ChooseRegionCTA\":\"Choose province or territory\",\"How\":\"Diagnosed with COVID-19?\",\"CTA\":\"What to do next\",\"HowUrl\":\"https://www.canada.ca/\",\"ExposedHelpCTA\":\"What to do if you’re exposed\",\"DiagnosedView\":{\"Title\":\"Thanks for helping fight COVID-19\",\"Body1\":{\"One\":\"Tomorrow, you’ll get your last notification asking permission to upload any new random codes your phone sent out.\",\"Other\":\"For the next {number} days, you’ll get a daily notification asking permission to share any new exposures your phone recorded.\"},\"Body2\":\"You do not have to agree, even if you agreed the day before.\",\"Body3\":\"If you’re worried about your symptoms, get medical care as soon as you can. Take care of yourself and your family during this time.\",\"Tip\":{\"Title\":\"TIP: \"}},\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DiagnosedShareView\":{\"Title\":\"Share new exposures\",\"Body1\":\"Thanks for helping slow the spread!\",\"Body2\":\"Reminder: \",\"Body3\":\"To help keep people safe, you need to share your new exposures every day, especially if you have to leave your home.\",\"ButtonCTA\":\"Share new exposures\"},\"DiagnosedShareUploadView\":{\"Title\":\"Not quite done\",\"Body1\":\"You entered your one-time key but you have not shared any exposures yet. People who were near you have not been notified yet.\",\"ButtonCTA\":\"Finish sharing exposures\"},\"BluetoothDisabled\":\"Bluetooth is off\",\"EnableBluetoothCTA\":\"You need to turn on Bluetooth in your phone’s settings so COVID Alert can work.\",\"TurnOnBluetooth\":\"Turn on Bluetooth\",\"EnUnauthorized\":{\"Title\":\"COVID Alert is off\",\"Body1\":\"From the button below, choose Exposure Notifications, then scroll down and turn on \\\"Share Exposure Information\\\".\",\"CTA\":\"Enable COVID Alert\"},\"EnDisabled\":{\"Title\":\"COVID Alert is off\",\"Body1\":\"You did not enable COVID Alert. The app needs your permission to work.\",\"CTA\":\"Enable COVID Alert\",\"AndroidTitle2\":\"Not sure about turning on Location?\",\"AndroidBody1\":\"You need to have Location turned on for COVID Alert to work. To use Bluetooth scanning, Android phones need Location setting on for all apps.\",\"AndroidBody2a\":\"Despite that, COVID Alert has \",\"AndroidBody2b\":\"no way of knowing where you are. \",\"AndroidBody2c\":\"You can check the app’s permissions in your phone’s settings. You’ll see it does not have permission for location services.\"},\"FrameworkUnavailable\":{\"Title\":\"Something went wrong\",\"Body\":\"To use COVID Alert, you need to update your phone’s operating system.\"},\"UnknownProblem\":{\"Title\":\"Something went wrong\",\"Body\":\"It looks like your phone cannot support COVID Alert.\",\"CTA\":\"Get Help\"},\"NoConnectivity\":\"COVID Alert is offline\",\"NoConnectivityDetailed\":\"COVID Alert only checks for exposure if your phone is connected to the Internet.\",\"ExternalLinkHint\":\"Opens in a new window\",\"LastCheckedMinutes\":{\"One\":\"Last checked {number} minute ago\",\"Other\":\"Last checked {number} minutes ago\"},\"LastCheckedHours\":{\"One\":\"Last checked {number} hour ago\",\"Other\":\"Last checked {number} hours ago\"},\"LastCheckedDays\":{\"One\":\"Last checked {number} day ago\",\"Other\":\"Last checked {number} days ago\"}},\"OverlayClosed\":{\"SystemStatus\":\"COVID Alert is \",\"SystemStatusOn\":\"active\",\"SystemStatusOff\":\"off\",\"NotificationStatus\":\"Notifications are \",\"NotificationStatusOff\":\"off\",\"TapPrompt\":\"Tap for menu\"},\"OverlayOpen\":{\"Version\":\"Version\",\"BluetoothCardBody\":\"You need to turn on Bluetooth in your phone’s settings so COVID Alert can work.\",\"BluetoothCardAction\":\"Turn on Bluetooth\",\"EnterCodeCardTitle\":\"Diagnosed with COVID-19?\",\"EnterCodeCardBody\":\"To let people know they’ve been exposed, you need a one-time key. We’ll also ask for details to help narrow down when you were likely most contagious.\",\"EnterCodeCardAction\":\"Enter your one-time key\",\"EnterCodeCardTitleDiagnosed\":\"You’re helping stop COVID-19\",\"EnterCodeCardBodyDiagnosed\":\"Thank you for doing your part to slow the spread. \",\"EnterCodeCardDiagnosedCountdown\":{\"One\":\"{number} day left!\",\"Other\":\"{number} days left!\"},\"CodeNotShared\":{\"Title\":\"Help slow the spread\",\"Body\":\"You entered your one-time key but you have not shared any exposures yet.\",\"CTA\":\"Finish sharing exposures\"},\"NotificationCardStatus\":\"Notifications are off\",\"NotificationCardStatusOff\":\"OFF\",\"NotificationCardBody\":\"You will not receive notifications if you’ve been near anyone who has reported a COVID-19 diagnosis. Finding out potential exposures right away is essential to slowing the spread of COVID-19.\",\"NotificationCardAction\":\"Turn on notifications\",\"ExposureNotificationCardStatus\":\"Exposure notifications are \",\"ExposureNotificationCardBody\":\"COVID Alert cannot notify you as soon as it learns of a possible exposure. Finding out right away is essential to slowing the spread of COVID-19.\",\"ExposureNotificationCardAction\":\"Turn on exposure notifications\",\"NoConnectivityCardAction\":\"COVID Alert is offline\",\"NoConnectivityCardBody\":\"Connect to the Internet to notify people or check for exposures.\",\"EnUnauthorizedCardAction\":\"Enable COVID Alert\",\"EnUnauthorizedCardBody\":\"Choose Exposure Notifications, then scroll down and turn on \\\"Share Exposure Information\\\".\",\"ShareDiagnosisCode\":{\"CtaReport\":\"Enter your one-time key\",\"CtaFinish\":\"Finish sharing exposures\"}},\"Info\":{\"CheckSymptoms\":\"Check symptoms\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/en\",\"TellAFriend\":\"Share app\",\"Support\":\"Support\",\"LearnMore\":\"How it works\",\"Help\":\"Help\",\"HelpUrl\":\"https://www.canada.ca/en/public-health/services/diseases/coronavirus-disease-covid-19/covid-alert/help.html\",\"GetCode\":\"Get a one-time key\",\"YourVisits\":\"Your Visits\",\"YourRecentExposures\":\"Your recent exposures\",\"ChangeLanguage\":\"Change language\",\"ToggleCovidAlert\":{\"TurnOff\":\"Turn off COVID Alert\",\"TurnOn\":\"Turn on COVID Alert\",\"Confirm\":{\"Title\":\"Turn off COVID Alert?\",\"Body\":\"Make sure you’re wearing full medical PPE (personal protective equipment) or you’re only with your household.  The app will not be able to record exposures.\",\"Cancel\":\"Keep it on\",\"Accept\":\"Turn it off\"}},\"ChangeRegion\":\"Change province or territory\",\"Privacy\":\"Privacy\",\"PrivacyUrl\":\"https://www.canada.ca/en/public-health/services/diseases/coronavirus-disease-covid-19/covid-alert/privacy-policy.html\",\"SettingsTitle\":\"Settings\",\"InformationTitle\":\"About COVID Alert\"},\"Landing\":{\"En\":\"English\",\"Fr\":\"Français\",\"AltText\":\"Illustration of 8 people enjoying a public space. There’s a person with a walking cane, a person in a wheelchair, a person wearing a hijab, several people of colour, and a representation of gender diversity.\",\"CanadaAltText\":\"Symbol of the Government of Canada\"},\"LanguageSelect\":{\"Title\":\"Language\",\"En\":\"English (Canada)\",\"Fr\":\"Français (Canada)\",\"EnShort\":\"English\",\"FrShort\":\"Français\",\"Close\":\"Close\"},\"Tutorial\":{\"step-1Title\":\"How COVID Alert works\",\"step-1AltText\":\"A hand-drawn illustration of a mobile phone in a pocket, sending and receiving signals.\",\"step-1\":\"The app runs in the background and will not interrupt your activities.\\n\\nWhenever you’re near someone else with COVID Alert, both phones exchange random codes every 5 minutes.\\n\\nThe random codes change often and cannot be used to identify you. \",\"step-2Title\":\"What’s an exposure?\",\"step-2AltText\":\"A hand-drawn illustration of 2 mobile phones side-by-side with a timer clock connecting both phones.\",\"step-2\":\"The app estimates how near people are by the strength of Bluetooth signals.\\n\\nThe app will record an exposure if the contact:\\n• Is closer than 2 metres.\\n• Lasts at least 15 minutes.\\n• Happens during the time the person was likely most contagious.\",\"step-3Title\":\"Getting a positive test\",\"step-3AltText\":\"A hand-drawn illustration of a mobile phone uploading a numeric code to a cloud.\",\"step-3\":\"If a person with the app is diagnosed with COVID-19, they can choose to upload the random codes their phone sent.\\n\\nThey can also give details to help narrow down when they were likely most contagious. Those details tell the phone which random codes to upload.\\n\\nThe random codes go into a central server. No other information is sent to the server.\",\"step-4Title\":\"Looking for exposures\",\"step-4AltText\":\"A hand-drawn illustration of a mobile phone uploading a numeric code to a cloud.\",\"step-4\":\"Every day, whenever it has an Internet connection, your phone will get a list of the random codes from people who reported a diagnosis.\\n\\nIf it finds codes that match, the app notifies you that you’ve been exposed and explains what to do next.\",\"ActionBack\":\"Back\",\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"Close\":\"Close\"},\"QRCodeOnboarding\":{\"step-1Title\":\"How QR code notifications work\",\"step-1AltText\":\"-\",\"step-1\":\"Public health identifies when a place has an exposure risk.\\n\\nEvery day, your app downloads a list of those places and times.\\n\\nIf there’s a match with one of your visits, the app notifies you and explains what to do next.\",\"step-2Title\":\"Your privacy is protected\",\"step-2AltText\":\"-\",\"step-2\":\"The app saves your visits on your phone, including the dates and times you were there. **Your visits and exposures are not shared** with anyone.\\n\\nYou can get to your recent visits from the menu. If you need to, you can delete any visit.\",\"step-3Title\":\"Help make contact tracing easier\",\"step-3AltText\":\"-\",\"step-3\":\"Next time you’re out and about, look for COVID Alert posters with QR codes.\\n\\nStart adding your visits and get notified of possible exposures from them.\",\"step-3-banner\":\"Right now, posters with QR codes are only available in _____ as part of a small-scale pilot project.\",\"ActionBack\":\"Back\",\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"Close\":\"Close\"},\"Onboarding\":{\"Step\":\"Step\",\"Of\":\"of\",\"ActionNext\":\"Next\",\"ActionEndSkip\":\"Skip\",\"ActionEnd\":\"Done\",\"ActionBack\":\"Back\",\"Start\":{\"Title\":\"Together, let’s stop the spread of COVID-19\",\"ImageAltText\":\"A hand-drawn illustration with lots of different people in it. Dotted lines show their paths crossing and leading to a smartphone below.\",\"Body1\":\"COVID Alert helps us break the cycle of infection. The app can let people know of possible exposures before any symptoms appear. \",\"Body2\":\"That way, we can take care of ourselves and protect our communities.\"},\"WhatItsNot\":{\"Title\":\"What COVID Alert does not do\",\"ImageAltText\":\"A hand-drawn illustration with many houses and a high-rise building on a curving street. Small message bubbles appear above each home.\",\"Body1\":\"The app will not tell you in the moment if you’re currently near someone who’s been diagnosed.\",\"Body2\":\"It will not tell you about outbreaks in your community.\"},\"Anonymous\":{\"Title\":\"Your privacy is protected\",\"ImageAltText\":\"A hand-drawn illustration of shaded glasses and hat to go incognito. Above the hat, a location pin icon is crossed out.\",\"Body1\":\"COVID Alert **does not use GPS** or track your location.\",\"Body2\":\"It has **no way of knowing**:\",\"Bullet1\":\"Your location.\",\"Bullet2\":\"Your name or address.\",\"Bullet3\":\"Your phone’s contacts.\",\"Bullet4\":\"Your health information, except what you decide to share.\",\"Bullet5\":\"The health information of anyone you’re near.\"},\"HowItWorks\":{\"Title\":\"How it works\",\"ImageAltText\":\"A hand-drawn illustration of 3 mobile phones sending out random codes via bluetooth.\",\"Body1\":\"The app uses Bluetooth to exchange random codes with nearby phones.\",\"Body2\":\"Every day, it checks a list of random codes from people who tell the app they tested positive.\",\"Body3\":\"If you’ve been near one of those people within the last 14 days, you’ll get a notification.\",\"HowItWorksCTA\":\"Learn more about how it works\"},\"PartOf\":{\"Title\":\"One part of public health\",\"ImageAltText\":\"A hand-drawn illustration of different ways of fighting COVID-19, including hand sanitizer, handwashing, face mask and the COVID Alert app.\",\"Body1\":\"COVID Alert is just one part of the public health effort to stop the spread of COVID-19.\",\"Body2\":\"Follow all public health guidelines in your area.\",\"Body3\":\"COVID Alert does not replace medical advice. If you get sick, contact your doctor or other healthcare provider.\"},\"Permissions\":{\"Title\":\"The app will now ask your permission\",\"ImageAltText\":\"A hand-drawn illustration of a mobile phone displaying a dialogue box with a red X and a green check mark on the screen.\",\"Body1\":\"**Allow the app** to start logging random codes or “random IDs” when you’re near other phones. The app will access the date, duration and signal strength related to the random codes, but they never leave your phone.\",\"Body2\":\"You’ll also need to **let the app send you notifications**.\",\"PrivacyButtonCTA\":\"Learn more about privacy\"}},\"Privacy\":{\"Title\":\"Privacy policy\",\"Close\":\"Close\"},\"ThankYou\":{\"Title\":\"Thank you for helping\",\"Body\":\"Exposure notifications are on. You will receive a message if COVID Alert detects that you’ve been near someone who has reported a COVID-19 diagnosis.\",\"Dismiss\":\"Dismiss\"},\"DataUpload\":{\"Cancel\":\"Cancel\",\"Close\":\"Close\",\"Back\":\"Back\",\"InfoSmall\":\"Your random codes will not be uploaded unless you give permission in the next step.\",\"ShareToast\":\"Random codes uploaded successfully\",\"StepXofY\":\"Step {x} of {y}\",\"Today\":\"Today\",\"Yesterday\":\"Yesterday\",\"Earlier\":\"Even earlier\",\"Step0\":{\"Title\":\"Notify people they’ve been exposed\",\"List\":{\"1a\":\"1. Enter the one-time key \",\"1b\":\"you got when you were diagnosed.\",\"2a\":\"2. Give the app details \",\"2b\":\"to help narrow down when you were likely most contagious. You can choose not to give any details.\",\"3a\":\"3. Allow your phone to share \",\"3b\":\"your random codes. These codes notify people you’ve been near.\"},\"Body1\":\"Exposed people will not get any information about you \",\"Body2\":\"or the time you were near them. They will only get a notification that they were exposed recently.\",\"CTA\":\"Next\",\"NoCode\":\"Need a one-time key?\"},\"IntermediateStep\":{\"Title\":\"Finish sharing exposures\",\"List\":{\"1a\":\"1. \",\"1b\":\"Done.\",\"2a\":\"2. Give the app details \",\"2b\":\"to help narrow down when you were likely most contagious. You can choose not to give any details.\",\"3a\":\"3. Allow your phone to share \",\"3b\":\"your random codes. These codes notify people you’ve been near.\"},\"Body1\":\"Nobody will get any information about you \",\"Body2\":\"or the time you were near them. They will only get a notification that they were exposed recently.\",\"CTA\":\"Next\"},\"Step2\":{\"Title1\":\"Are you feeling sick?\",\"Body1\":\"Symptoms could include:\",\"Symptoms\":{\"1\":\"Cough or breathing problems.\",\"2\":\"Fever or chills.\",\"3\":\"Feeling really tired or weak.\",\"4\":\"Sore muscles or headache.\",\"5\":\"New loss of smell or taste.\",\"6\":\"Stomach problems.\"},\"Option1\":\"Yes, you have or had at least 1 symptom within the last 14 days.\",\"Option2\":\"No, you’ve not had any symptoms.\",\"Option3\":\"You're not sure.\",\"Option4\":\"You prefer not to say.\",\"Title2\":\"Why we’re asking\",\"Body2\":\"This information helps figure out when you were likely most contagious. It tells the app which exposures to share so only people who were near you during that time will be notified. It will never be sent or stored outside your phone.\",\"CTA\":\"Next\"},\"FormView\":{\"Title\":\"Enter your one-time key\",\"Body\":\"Enter the key you got when you were diagnosed. To prevent false notifications, you can only get a key if you test positive for COVID-19.\",\"Action\":\"Submit key\",\"InputLabel\":\"COVID Alert key\"},\"OtkDiagnosedView\":{\"Title\":\"You've already entered the key\",\"Body\":\"You do not need to enter it again.\",\"CTA\":\"Next\"},\"ConsentView\":{\"Title\":\"Now you can upload your codes\",\"Body1\":\"Your phone will ask your permission to upload your most recent random codes or “random IDs”.\",\"Body2a\":\"Nobody will get any information about you \",\"Body2b\":\"or the time you were near them.\",\"Body3\":\"You can help slow the spread of COVID-19 if you agree.\",\"Action\":\"Agree\"},\"TekUploadNoDate\":{\"Title\":\"Notify people you've been near\",\"Body\":\"You did not give a date to help narrow down when you were likely most contagious.\\n\\nShare your exposures now to notify the people you were near in the last 14 days.\",\"CTA\":\"Share exposures\"},\"SymptomOnsetDate\":{\"Title1\":\"Date your symptoms started\",\"Title2\":\"Why we’re asking\",\"Body1\":\"This date helps figure out when you were likely most contagious. It tells the app which exposures to share so only people who were near you during that time will be notified.\\n\\nThe date does not get sent or stored anywhere outside your phone.\",\"CTA\":\"Share exposures\",\"CTA2\":\"Prefer not to say\"},\"TestDate\":{\"Title1\":\"Your test date\",\"Body1\":\"Enter the date you got your COVID-19 test.\",\"Title2\":\"Why we’re asking\",\"Body2\":\"Your test date helps figure out when you were likely most contagious. It tells the app which exposures to share so only people who were near you then will be notified.\\n\\nThe date does not get sent or stored anywhere outside your phone.\",\"CTA\":\"Share exposures\",\"CTA2\":\"Prefer not to say\"},\"NoCode\":{\"NoRegion\":{\"Title\":\"You skipped choosing province or territory\",\"Body\":\"In some provinces and territories, people cannot yet report a COVID-19 diagnosis through this app.\\n\\nTo find out if you can get a one-time key, you need to choose your area.\",\"ChooseRegionCTA\":\"Choose province or territory\"},\"RegionNotCovered\":{\"Title\":\"No reporting yet in your area\",\"Body\":\"People in your area cannot yet get a one-time key.\\n\\nYou will not be able to let people know they were exposed. But you can still self-isolate for 14 days, and:\",\"Body2\":\"Keep your community safe.\",\"Body3\":\"Protect people and pets you live with.\",\"Body4\":\"Take care of yourself.\"}}},\"Notification\":{\"ExposedMessageTitle\":\"You’ve been exposed\",\"ExposedMessageBody\":\"You’ve had close contact with someone who reported a COVID-19 diagnosis through the app. Learn more about what to do next.\",\"OffMessageTitle\":\"COVID Alert is off\",\"OffMessageBody\":\"Turn on COVID Alert to get notified if you’ve been near someone who has reported a COVID-19 diagnosis.\",\"DailyUploadNotificationTitle\":\"Upload new random codes\",\"DailyUploadNotificationBody\":\"Help slow the spread of COVID-19 and upload your new random codes.\",\"ExposureChecksTitle\":\"COVID Alert\",\"ExposureChecksBody\":\"Checking for exposures\",\"ExposureChecksAndroidChannelName\":\"Checking for exposures\",\"PausedMessageTitle\":\"You turned off the app\",\"PausedMessageBody\":\"The app is not recording exposures because you turned it off.\",\"OTKNotSharedTitle\":\"Not quite done\",\"OTKNotSharedBody\":\"You entered your one-time key but you have not shared any exposures yet. People who were near you have not been notified yet.\",\"AndroidChannelName\":\"COVID Alert\",\"OutbreakMessageTitle\":\"You have been exposed to an Outbreak\",\"OutbreakMessageBody\":\"There was an outbreak at one of the locations you scanned\"},\"Partners\":{\"Label\":\"Developed in partnership with\"},\"BottomSheet\":{\"Collapse\":\"Close\",\"OnStatus\":\"COVID Alert is on, Tap for more information\",\"OffStatus\":\"COVID Alert is off, Tap for more information\"},\"RegionLanding\":{\"Title\":\"Choose your province or territory\",\"Body1\":\"Provincial and territorial governments are working to support COVID Alert across Canada.\",\"Body2\":\"In some places, people cannot yet report a COVID-19 diagnosis through this app.\",\"RegionSelectBtn\":\"Choose province or territory\"},\"RegionPicker\":{\"Title\":\"Where do you live? (optional)\",\"SettingsTitle\":\"Province or territory\",\"Close\":\"Close\",\"Body\":\"Choose your region to check if you can report a diagnosis through this app.\",\"Skip\":\"Skip\",\"GetStarted\":\"Get started\",\"Optional\":\"* Optional\",\"AB\":\"Alberta\",\"BC\":\"British Columbia\",\"MB\":\"Manitoba\",\"NB\":\"New Brunswick\",\"NL\":\"Newfoundland and Labrador\",\"NT\":\"Northwest Territories\",\"NS\":\"Nova Scotia\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Prince Edward Island\",\"QC\":\"Quebec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\",\"None\":\"Prefer not to say\",\"ExposedHelpLinkNote\":{\"AB\":\"\",\"BC\":\"\",\"SK\":\"\"}},\"RegionPickerNoPT\":{\"Exposed\":{\"Title\":\"Find your area’s advice\",\"Body\":\"The steps you need to take when you’re exposed depend on the province or territory you live in.\"}},\"A11yList\":{\"Start\":\"List start\",\"End\":\"List end\",\"Bullet\":\"Bullet\"},\"MenuButton\":\"Menu\",\"QRCode\":{\"CTA\":\"Add a visit\",\"Reader\":{\"Title\":\"Point the camera at the QR code\",\"Body\":\"Subscribe to a place to receive notification if other people visit this place at a similar time as you.\",\"Back\":\"Back\",\"Learn\":\"Learn more about QR codes and COVID Alert\"},\"CameraPermissions\":{\"Title\":\"Give access to camera\",\"Body\":\"To scan a place, the app needs to use your camera.\",\"Title2\":\"Not sure about giving access?\",\"Body2\":\"You can also use your phone’s camera directly if you prefer not to let the app access your camera.\",\"CTA\":\"Next\",\"Cancel\":\"Not now\"},\"CameraPermissionDenied\":{\"Title\":\"Access denied\",\"Body\":\"You said no to COVID Alert using your phone’s camera.\\n\\nTo scan a place, you’ll need to update your phone’s settings to let COVID Alert  use the camera.\\n\\nYou can also use your phone’s camera directly or a QR scanning app.\",\"CTA\":\"Allow camera access in settings\",\"CTA2\":\"About adding a visit\"},\"CameraPermissionAlert\":{\"Title\":\"« COVID Alert » Would Like to Access the Camera\",\"Subtitle\":\"COVID Alert needs access to your camera so it can read the QR code.\",\"Allow\":\"Allow\",\"DontAllow\":\"Don’t Allow\"},\"LearnAboutQRScan\":{\"Title\":\"How to add a visit\",\"Body1\":\"1. Give COVID Alert access to your camera so it can read the QR code.\",\"Body2\":\"2. Point your phone at the code.\"},\"Error\":{\"Title\":\"QR code not recognized\",\"Body\":\"You may not have scanned an official COVID Alert QR code.\\n\\nAsk someone for the official COVID Alert poster.\",\"CTA\":\"Try Again\",\"CTA2\":\"Back to Home\"},\"CheckInView\":{\"Title\":\"Thank you for adding a visit\",\"Body1\":\"You’ll only be notified if someone from today tests positive. \",\"Body2\":\"Your visits are not shared \",\"Body3\":\"with anyone.\\n\\nRemember to scan each time you visit.\",\"Body4\":\"You can delete this visit any time by going into Your visits.\",\"CTA1\":\"Back to home\",\"CTA2\":\"Your visits\",\"MoreInfo\":\"Learn more about QR codes and COVID alert\"},\"Exposed\":{\"Title\":\"You've been Exposed\"},\"CancelCheckIn\":{\"Title\":\"Scan cancelled\"},\"ScanAPlace\":{\"Title\":\"Add a visit\",\"Body\":\"Get notified if someone from today tests positive later.\\n\\nThe app saves your visits on your phone, including the dates and times you were there.\\n\\n**Your visits are not shared** with anyone.\",\"CTA\":\"Next\",\"CTA2\":\"About adding a visit\"},\"ProximityExposure\":\"Get tested\",\"OutbreakExposed\":{\"Title\":\"Possible exposure at a visit\",\"Body\":\"Someone tested positive from one of your visits.\",\"CTA\":\"Negative test?\",\"SelfIsolate\":{\"Title\":\"Isolate and get tested\",\"Body\":\"You’re a high-risk contact. If you have symptoms, get tested right away. If you do not have symptoms, wait to get tested at least 7 days after the date of exposure.\\n\\nYou also need to isolate for 14 days from the date of exposure.\\n\\nIf you have questions, contact your local public health unit.\"},\"GetTested\":{\"Title\":\"Get tested\",\"Body\":\"You need to get tested as soon as possible.\\n\\nIn the meantime, avoid contact with vulnerable people.\\n\\nIf you have questions, contact your local public health unit.\"},\"SelfMonitor\":{\"Title\":\"Avoid people and self-monitor\",\"Body\":\"You need to avoid contact with vulnerable people and watch yourself carefully for symptoms. If you get symptoms, take a self-assessment.\\n\\nIf you have questions, contact your local public health unit.\"}}},\"PlacesLog\":{\"Title\":\"Your Visits\",\"Body1\":\"You’ll get notified if there’s a possible exposure from the following visits.\",\"Body2a\":\"Your visits are not shared \",\"Body2b\":\"with anyone. They’re only on your phone.\",\"DeleteBtnCTA\":\"Delete all visits\",\"DeleteIcon\":\"Delete\",\"NoVisit\":\"No visits yet\",\"Back\":\"Back\",\"Alert\":{\"Title\":\"Are you sure you want to delete this visit?\",\"TitleDeleteAll\":\"Are you sure you want to delete all visits?\",\"Subtitle\":\"If there’s a possible exposure at this place, you will not be notified.\",\"SubtitleDeleteAll\":\"If there’s a possible exposure at any of these places, you will not be notified.\",\"Cancel\":\"Cancel\",\"ConfirmDelete\":\"Yes, delete\",\"ConfirmDeleteAll\":\"Yes, delete all\"}},\"ExposureHistory\":{\"Title\":\"Your recent exposures\",\"Body\":\"This list shows your exposures from the last 14 days.\",\"NoExposures\":\"No exposures\"},\"Errors\":{\"Action\":\"OK\",\"TekUploadNoDate\":{\"Title\":\"\",\"Body\":\"You need to choose a date before you can share exposures.\"},\"TekUploadOffline\":{\"Title\":\"Lost connection\",\"Body\":\"Your random codes were not uploaded. Make sure your phone is connected to the Internet, through either data or WiFi.\"},\"TekUploadServer\":{\"Title\":\"Try again later\",\"Body\":\"Something went wrong with the system.\"},\"TekUploadPermission\":{\"Title\":\"Random codes could not be uploaded\",\"Body\":\"You did not give permission.\"},\"OtcCodeNotEntered\":{\"Title\":\"\",\"Body\":\"You need to enter your key before you can submit.\"},\"OtcUploadInvalidOneTimeCode\":{\"Title\":\"Incorrect key\",\"Body\":\"Your one-time key could not be recognized. Please try again.\"},\"OtcUploadTemporaryBan\":{\"Title\":\"Wait for 1 hour\",\"Body\":\"You’ve tried to enter a one-time key too many times.\"},\"OtcUploadDefault\":{\"Title\":\"Try again later\",\"Body\":\"Something went wrong with the system.\"},\"OtcUploadOffline\":{\"Title\":\"Lost connection\",\"Body\":\"Your key was not sent. Make sure your phone is connected to the Internet, through either data or WiFi.\"},\"ContentNoConnect\":{\"Title\":\"Cannot connect to information for your province or territory\",\"Body\":\"Try again later for this information.\"},\"DataSharingStep2\":{\"Title\":\"\",\"Body\":\"You need to choose an answer before going to the next step.\"},\"ApiNotConnected\":{\"Title\":\"Something went wrong\",\"Body\":\"Unfortunately, it looks like COVID Alert will not work on this phone. \\n\\nIf you have more than one profile on your phone, try installing the app in the main profile.\"}}},\"fr\":{\"Home\":{\"AppName\":\"Alerte COVID\",\"UserStopped\":{\"Title\":\"Vous avez désactivé l’application\",\"CTA\":\"Réactiver\"},\"NoExposureDetected\":{\"AllSetTitle\":\"Vous voilà prêt!\",\"RegionCovered\":{\"Title\":\"Aucune exposition détectée\",\"Body\":\"Vous n’avez été près d’aucune personne ayant signalé un diagnostic de COVID-19 avec cette application.\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"AllSetBody\":\"Votre téléphone utilise maintenant Bluetooth pour collecter les codes provenant des téléphones voisins.\"},\"NoRegion\":{\"Title\":\"Aucune exposition détectée\",\"Body\":\"Vous n’avez été près d’aucune personne ayant signalé un diagnostic de COVID-19 avec cette application.\\n\\nToutefois, dans certaines provinces et certains territoires, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\",\"AllSetBody\":\"Nous vous informerons dès que les personnes en dehors de l’Ontario peuvent signaler des diagnostics avec Alerte COVID.\\n\\nIl est tout de même utile de garder Alerte COVID activée.\\n\\nDe cette façon, vous pourrez être avisé dès que les personnes dont vous avez été près signalent un diagnostic.\"},\"RegionNotCovered\":{\"Title\":\"Pas encore de signalement dans votre région\",\"Body\":\"Les personnes de votre province ou territoire ne peuvent pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nIl est tout de même utile d’activer Alerte COVID. Quand les personnes pourront signaler un diagnostic, vous serez ainsi avisé si jamais vous avez été près d’elles.\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"AllSetBody\":\"Nous vous informerons dès que les personnes de votre région peuvent signaler un diagnostic avec Alerte COVID.\\n\\nIl est tout de même utile de garder Alerte COVID activée.\\n\\nDe cette façon, vous pourrez être avisé dès que les personnes dont vous avez été près signalent un diagnostic.\"}},\"ExposureDetected\":{\"Title\":\"Vous avez été exposé\",\"Body1\":\"Vous avez côtoyé au moins une personne qui a signalé un diagnostic de COVID-19 au moyen de l’application. Le contact a duré 15 minutes ou plus.\",\"Title2\":\"Et ensuite?\",\"Body2\":\"Vous pourriez avoir contracté le virus.\",\"Notification\":{\"Received\":\"Date de la notification\"},\"NegativeTest\":{\"CTA\":\"Résultat de test négatif?\"},\"Dismiss\":{\"Title\":\"Résultat négatif au test de dépistage?\",\"Body\":\"Vous pouvez mettre fin à l’alerte pour continuer à recevoir des notifications d’exposition. Aucune information d’exposition ne sera supprimée de votre téléphone.\",\"Body2\":\"Vous pouvez aussi mettre fin à l’alerte si vous suivez les consignes de votre région.\",\"CTA\":\"Mettre fin à l’alerte\",\"Confirm\":{\"Body\":\"Voulez-vous vraiment mettre fin à l’alerte?\",\"Accept\":\"Oui, mettre fin à l’alerte\",\"Cancel\":\"Annuler\"}}},\"ChooseRegionCTA\":\"Sélectionnez une province ou un territoire\",\"How\":\"Diagnosed with COVID-19?\",\"CTA\":\"Prochaines étapes\",\"HowUrl\":\"https://www.canada.ca/\",\"ExposedHelpCTA\":\"Que faire en cas d’exposition\",\"DiagnosedView\":{\"Title\":\"Merci d’aider à lutter contre la COVID\",\"Body1\":{\"One\":\"Demain, vous recevrez une dernière notification vous demandant l’autorisation de téléverser tout nouveau code aléatoire émis par votre téléphone.\",\"Other\":\"Au cours des {number} prochains jours, vous recevrez une notification quotidienne vous demandant l’autorisation de partager toute nouvelle exposition enregistrée par votre téléphone.\"},\"Body2\":\"Vous n’êtes pas tenu d’accepter, même si vous avez accepté le jour d’avant.\",\"Body3\":\"Si vos symptômes vous inquiètent, vous pouvez consulter un professionnel de la santé. Prenez soin de vous et de votre famille pendant ce temps.\",\"Tip\":{\"Title\":\"CONSEIL : \"}},\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DiagnosedShareView\":{\"Title\":\"Partager les nouvelles expositions\",\"Body1\":\"Merci d’aider à ralentir la propagation de la COVID-19!\",\"Body2\":\"Rappel : \",\"Body3\":\"Pour contribuer à la sécurité des autres, vous devez partager vos nouvelles expositions tous les jours, surtout si vous avez à sortir de la maison.\",\"ButtonCTA\":\"Partager les nouvelles expositions\"},\"DiagnosedShareUploadView\":{\"Title\":\"Pas tout à fait fini\",\"Body1\":\"Vous avez entré votre clé à usage unique, mais vous n’avez pas encore partagé vos expositions. Les personnes que vous avez côtoyées n’ont pas encore été notifiées.\",\"ButtonCTA\":\"Finir le partage d’expositions\"},\"BluetoothDisabled\":\"Bluetooth désactivé\",\"EnableBluetoothCTA\":\"Vous devez activer la fonction Bluetooth dans les paramètres de votre téléphone pour qu’Alerte COVID fonctionne.\",\"TurnOnBluetooth\":\"Activez Bluetooth\",\"EnUnauthorized\":{\"Title\":\"Alerte COVID est désactivée\",\"Body1\":\"Appuyez sur le bouton ci-dessous, puis sélectionnez Notifications d’exposition. Faites défiler l’écran et activez « Partager les informations sur les expositions ».\",\"CTA\":\"Activer Alerte COVID\"},\"EnDisabled\":{\"Title\":\"Alerte COVID est désactivée\",\"Body1\":\"Vous n’avez pas activé Alerte COVID. L’application a besoin de votre autorisation pour fonctionner.\",\"CTA\":\"Activer Alerte COVID\",\"AndroidTitle2\":\"Vous hésitez à activer la localisation?\",\"AndroidBody1\":\"Il faut activer la localisation pour qu’Alerte COVID fonctionne. Avec Android, la recherche Bluetooth n’est possible que si la localisation est activée pour toutes les applications.\",\"AndroidBody2a\":\"Malgré cela, Alerte COVID \",\"AndroidBody2b\":\"ne peut pas savoir où vous êtes. \",\"AndroidBody2c\":\"Vous pouvez vous en assurer en vérifiant les autorisations d’Alerte COVID dans les paramètres de votre téléphone. La localisation n’est pas autorisée.\"},\"FrameworkUnavailable\":{\"Title\":\"Une erreur s’est produite\",\"Body\":\"Pour utiliser Alerte COVID, vous devez mettre à jour le système d’exploitation de votre téléphone.\"},\"UnknownProblem\":{\"Title\":\"Une erreur s’est produite\",\"Body\":\"Il semble que votre téléphone ne puisse pas prendre en charge Alerte COVID.\",\"CTA\":\"Trouver de l’aide\"},\"NoConnectivity\":\"Alerte COVID est hors ligne\",\"NoConnectivityDetailed\":\"Alerte COVID vérifie s’il y a des expositions seulement quand votre téléphone est connecté à Internet.\",\"ExternalLinkHint\":\"Ouvre dans une nouvelle fenêtre\",\"LastCheckedMinutes\":{\"One\":\"Dernière vérification il y a {number} minute\",\"Other\":\"Dernière vérification il y a {number} minutes\"},\"LastCheckedHours\":{\"One\":\"Dernière vérification il y a {number} heure\",\"Other\":\"Dernière vérification il y a {number} heures\"},\"LastCheckedDays\":{\"One\":\"Dernière vérification il y a {number} jour\",\"Other\":\"Dernière vérification il y a {number} jours\"}},\"OverlayClosed\":{\"SystemStatus\":\"Alerte COVID est \",\"SystemStatusOn\":\"activée\",\"SystemStatusOff\":\"désactivée\",\"NotificationStatus\":\"Notifications \",\"NotificationStatusOff\":\"désactivées\",\"TapPrompt\":\"Appuyez pour voir le menu\"},\"OverlayOpen\":{\"Version\":\"Version \",\"BluetoothCardBody\":\"Vous devez l’activer dans les paramètres de votre téléphone pour qu’Alerte COVID fonctionne.\",\"BluetoothCardAction\":\"Activez Bluetooth\",\"EnterCodeCardTitle\":\"Avez-vous reçu un diagnostic de COVID-19?\",\"EnterCodeCardBody\":\"Vous devez avoir une clé à usage unique pour aviser les personnes exposées. Nous vous demanderons aussi quelques précisions pour déterminer la période où vous étiez probablement le plus contagieux.\",\"EnterCodeCardAction\":\"Entrer votre clé\",\"EnterCodeCardTitleDiagnosed\":\"Vous contribuez à freiner la COVID\",\"EnterCodeCardBodyDiagnosed\":\"Merci de faire votre part pour freiner la propagation de la COVID-19. \",\"EnterCodeCardDiagnosedCountdown\":{\"One\":\"Il vous reste {number} jour!\",\"Other\":\"Il vous reste {number} jours!\"},\"CodeNotShared\":{\"Title\":\"Aidez à ralentir la propagation de la COVID-19 \",\"Body\":\"Vous avez entré votre clé à usage unique, mais vous n’avez pas encore partagé vos expositions.\",\"CTA\":\"Finir le partage d’expositions\"},\"NotificationCardStatus\":\"Les notifications sont désactivées\",\"NotificationCardStatusOff\":\"DÉSACTIVÉES\",\"NotificationCardBody\":\"Alerte COVID ne peut pas vous aviser dès qu’il découvre une exposition potentielle. Le fait d’être notifié rapidement est essentiel pour ralentir la propagation de la COVID-19.\",\"NotificationCardAction\":\"Activer les notifications\",\"ExposureNotificationCardStatus\":\"Les notifications d’exposition à la COVID-19 sont \",\"ExposureNotificationCardBody\":\"Alerte COVID ne peut pas vous aviser de la découverte d’une exposition potentielle. Le fait d’être notifié rapidement est essentiel pour ralentir la propagation de la COVID-19.\",\"ExposureNotificationCardAction\":\"Activez les notifications d’exposition\",\"NoConnectivityCardAction\":\"Alerte COVID est hors ligne\",\"NoConnectivityCardBody\":\"Connectez-vous à Internet pour aviser les autres personnes ou pour vérifier si vous avez été exposé.\",\"EnUnauthorizedCardAction\":\"Activer Alerte COVID\",\"EnUnauthorizedCardBody\":\"Sélectionnez Notifications d’exposition. Faites défiler l’écran et activez « Partager les informations sur les expositions ».\",\"ShareDiagnosisCode\":{\"CtaReport\":\"Entrer une clé à usage unique\",\"CtaFinish\":\"Finir le partage d’expositions\"}},\"Info\":{\"CheckSymptoms\":\"Vérifier vos symptômes\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/fr\",\"TellAFriend\":\"Partager l’application\",\"Support\":\"Aide\",\"LearnMore\":\"Fonctionnement\",\"Help\":\"Aide et questions fréquentes\",\"HelpUrl\":\"https://www.canada.ca/fr/sante-publique/services/maladies/maladie-coronavirus-covid-19/alerte-covid/aide.html\",\"GetCode\":\"Obtenir une clé à usage unique\",\"YourVisits\":\"Vos visites\",\"YourRecentExposures\":\"Vos expositions récentes\",\"ChangeLanguage\":\"Modifier la langue\",\"ToggleCovidAlert\":{\"TurnOff\":\"Désactiver Alerte COVID\",\"TurnOn\":\"Réactiver Alerte COVID\",\"Confirm\":{\"Title\":\"Désactiver Alerte COVID?\",\"Body\":\"Veillez à porter un équipement de protection individuelle (ÉPI) médical ou à ne côtoyer que les membres de votre ménage. Une fois désactivée, l’application ne pourra plus enregistrer d’expositions.\",\"Cancel\":\"Ne pas désactiver\",\"Accept\":\"Désactiver\"}},\"ChangeRegion\":\"Modifier la province ou le territoire\",\"Privacy\":\"Confidentialité\",\"PrivacyUrl\":\"https://www.canada.ca/fr/sante-publique/services/maladies/maladie-coronavirus-covid-19/alerte-covid/politique-confidentialite.html\",\"SettingsTitle\":\"Paramètres\",\"InformationTitle\":\"À propos d’Alerte COVID\"},\"Landing\":{\"En\":\"English\",\"Fr\":\"Français\",\"AltText\":\"Illustration de 8 personnes profitant de l’espace public. Il y a une personne avec une canne blanche, une autre en fauteuil roulant, une autre avec un hijab, plusieurs personnes de couleur, et une représentation de la diversité de genre.\",\"CanadaAltText\":\"Symbole du gouvernement du Canada\"},\"LanguageSelect\":{\"Title\":\"Langue\",\"En\":\"English (Canada)\",\"Fr\":\"Français (Canada)\",\"EnShort\":\"English\",\"FrShort\":\"Français\",\"Close\":\"Fermer\"},\"Tutorial\":{\"step-1Title\":\"Fonctionnement d’Alerte COVID\",\"step-1AltText\":\"Dessin à la main d’un téléphone mobile dans une poche, qui envoie et reçoit des signaux.\",\"step-1\":\"L’application s’exécute en arrière-plan sans interrompre vos activités.\\n\\nQuand vous serez près d’une personne ayant Alerte COVID, vos téléphones s’échangeront des codes aléatoires toutes les 5 minutes.\\n\\nCes codes changent souvent et ne peuvent pas servir à vous identifier.\",\"step-2Title\":\"Qu’est-ce qu’une exposition?\",\"step-2AltText\":\"Dessin à la main de 2 téléphones mobiles côte à côte avec un chronomètre qui relie les deux téléphones.\",\"step-2\":\"L’application estime la distance entre deux personnes grâce à la puissance des signaux Bluetooth.\\n\\nPour que l’application enregistre une exposition, un contact doit :\\n• Avoir lieu à moins de 2 mètres.\\n• Durer au moins 15 minutes.\\n• Avoir lieu quand l’autre personne est probablement le plus contagieuse.\",\"step-3Title\":\"En cas de test positif\",\"step-3AltText\":\"Dessin à la main d’un téléphone mobile téléversant un code chiffré vers un nuage.\",\"step-3\":\"Si une personne ayant l’application reçoit un diagnostic de COVID-19, elle peut décider de téléverser les codes aléatoires émis par son téléphone.\\n\\nLa personne peut aussi fournir quelques précisions pour déterminer la période où elle était probablement le plus contagieuse. Cette période indique au téléphone les codes aléatoires qu’il faut téléverser.\\n\\nLes codes aléatoires sont envoyés vers un serveur central. Celui-ci ne reçoit aucun autre renseignement.\",\"step-4Title\":\"Découverte des expositions\",\"step-4AltText\":\"Dessin à la main d’un téléphone mobile émettant une liste de codes chiffrés. L’icône de la loupe de recherche met en évidence l’un des codes de la liste.\",\"step-4\":\"Chaque jour, quand votre téléphone a une connexion Internet, il obtient une liste de codes aléatoires émis par les personnes qui ont signalé un diagnostic.\\n\\nSi l’application reconnaît des codes dans cette liste, elle vous dira que vous avez été exposé et vous expliquera quoi faire ensuite.\",\"ActionBack\":\"Précédent\",\"ActionNext\":\"Suivant\",\"ActionEnd\":\"Terminé\",\"Close\":\"Fermer\"},\"QRCodeOnboarding\":{\"step-1Title\":\"Fonctionnement des notifications par code QR\",\"step-1AltText\":\"-\",\"step-1\":\"Le lieu et le moment où il y a un risque d’exposition est déterminé par les autorités de santé publique.\\n\\nChaque jour, Alerte COVID télécharge une liste de ces lieux et moments.\\n\\nS’il y a correspondance avec l’une de vos visites, l’application vous enverra une notification et vous expliquera quoi faire.\",\"step-2Title\":\"Votre vie privée est protégée\",\"step-2AltText\":\"-\",\"step-2\":\"L’application enregistre la date, l’heure et le lieu de vos visites sur votre téléphone. Vos visites et vos expositions ne sont communiquées à personne.\\n\\nVous pouvez consulter la liste de vos visites récentes à partir du menu. Au besoin, vous pouvez aussi supprimer des visites.\",\"step-3Title\":\"Facilitons la recherche de contacts\",\"step-3AltText\":\"-\",\"step-3\":\"La prochaine fois que vous sortez, repérez les affiches Alerte COVID avec des codes QR.\\n\\nPuis, commencez à ajouter vos visites et soyez avisé s’il y a eu risque d’exposition pendant celles-ci.\",\"step-3-banner\":\"Pour l’instant, les affiches à code QR ne sont présentes qu’à ________ pour un projet-pilote à petite échelle.\",\"ActionBack\":\"Back\",\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"Close\":\"Close\"},\"Onboarding\":{\"Step\":\"Étape\",\"Of\":\"de\",\"ActionNext\":\"Suivant\",\"ActionEndSkip\":\"Ignorer\",\"ActionEnd\":\"Terminé\",\"ActionBack\":\"Précédent\",\"Start\":{\"Title\":\"Unissons-nous pour freiner la propagation de la COVID-19\",\"ImageAltText\":\"Dessin à la main avec plusieurs personnes différentes dedans. Leurs chemins en lignes pointillées se croisent et mènent à un téléphone intelligent plus bas.\",\"Body1\":\"Alerte COVID nous aide à briser le cycle d’infection. L’application permet d’aviser les personnes en cas d’exposition potentielle avant même que des symptômes apparaissent. \",\"Body2\":\"Nous pouvons ainsi prendre soin de nous et protéger nos communautés.\"},\"WhatItsNot\":{\"Title\":\"Ce qu’Alerte COVID ne fait pas\",\"ImageAltText\":\"Dessin à la main d’une rue sinueuse avec plusieurs maisons et un immeuble en hauteur. De petites infobulles apparaissent au-dessus de chaque bâtiment.\",\"Body1\":\"L’application ne vous dira pas sur-le-champ que vous êtes près d’une personne ayant eu un diagnostic.\",\"Body2\":\"Elle ne vous renseignera pas sur les éclosions de COVID-19 dans votre collectivité.\"},\"Anonymous\":{\"Title\":\"Votre vie privée est protégée\",\"ImageAltText\":\"Dessin à la main d’une paire de lunettes de soleil et d’un chapeau pour passer incognito. Au-dessus du chapeau, une icône de marqueur de géolocalisation est biffée.\",\"Body1\":\"Alerte COVID **n’utilise pas le GPS** et ne suit pas votre emplacement.\",\"Body2\":\"L’application **n’a aucun moyen** de connaître :\",\"Bullet1\":\"Votre emplacement.\",\"Bullet2\":\"Votre nom ou votre adresse.\",\"Bullet3\":\"Les contacts de votre téléphone.\",\"Bullet4\":\"Vos renseignements de santé, sauf ceux que vous acceptez de fournir.\",\"Bullet5\":\"Les renseignements de santé des personnes à proximité.\"},\"HowItWorks\":{\"Title\":\"Fonctionnement\",\"ImageAltText\":\"Dessin à la main de 3 téléphones mobiles émettant des codes aléatoires en utilisant Bluetooth.\",\"Body1\":\"L’application utilise Bluetooth pour échanger des codes aléatoires avec les téléphones à proximité.\",\"Body2\":\"Chaque jour, elle vérifie une liste de codes aléatoires provenant des personnes ayant signalé un test positif à l’application.\",\"Body3\":\"Si vous avez côtoyé l’une de ces personnes au cours des 14 derniers jours, vous recevrez une notification.\",\"HowItWorksCTA\":\"En savoir plus sur le fonctionnement\"},\"PartOf\":{\"Title\":\"L’un des efforts de santé publique\",\"ImageAltText\":\"Dessin à la main de divers moyens de lutter contre la COVID-19, dont un désinfectant pour les mains, un couvre-visage, des mains avec du savon et l’application Alerte COIVID.\",\"Body1\":\"Alerte COVID n’est qu’une partie des efforts de santé publique visant à freiner la propagation de la COVID-19.\",\"Body2\":\"Continuez à suivre les recommandations de santé publique de votre région.\",\"Body3\":\"Alerte COVID ne remplace pas un avis médical. Si vous tombez malade, contactez votre médecin ou un professionnel de la santé.\"},\"Permissions\":{\"Title\":\"L’application demandera votre autorisation\",\"ImageAltText\":\"Dessin à la main d’un téléphone mobile dont l’écran affiche une fenêtre de dialogue avec un X rouge et un crochet vert.\",\"Body1\":\"**Permettez à l’application** de collecter des codes aléatoires quand vous êtes près d’autres téléphones. L’application connaîtra la date, la durée et la puissance de signal associées à ces codes, mais ces données resteront sur votre téléphone. \",\"Body2\":\"Vous devrez aussi **lui permettre de vous envoyer des notifications**.\",\"PrivacyButtonCTA\":\"En savoir plus sur la confidentialité\"}},\"Privacy\":{\"Title\":\"Politique de confidentialité\",\"Close\":\"Fermer\"},\"ThankYou\":{\"Title\":\"Merci d’aider\",\"Body\":\"Les notifications d’exposition sont activées. Si Alerte COVID détecte que vous avez possiblement été exposé(e) à la COVID-19, vous recevrez une notification.\",\"Dismiss\":\"OK\"},\"DataUpload\":{\"Cancel\":\"Annuler\",\"Close\":\"Fermer\",\"Back\":\"Précédent\",\"InfoSmall\":\"Vos identifiants aléatoires ne seront pas partagés, à moins que vous donniez votre permission à l’étape suivante.\",\"ShareToast\":\"Identifiants aléatoires partagés avec succès\",\"StepXofY\":\"Étape {x} de {y}\",\"Today\":\"Aujourd’hui\",\"Yesterday\":\"Hier\",\"Earlier\":\"Il y a plus longtemps\",\"Step0\":{\"Title\":\"Aviser les personnes d’une exposition\",\"List\":{\"1a\":\"1. Vous entrez la clé à usage unique \",\"1b\":\"reçue lors de votre diagnostic.\",\"2a\":\"2. Vous fournissez quelques précisions \",\"2b\":\"qui aideront l’application à déterminer la période où vous étiez probablement le plus contagieux. Vous n’êtes pas obligé de le faire.\",\"3a\":\"3. Vous autorisez votre téléphone \",\"3b\":\"à partager vos codes aléatoires. Ceux-ci permettent d’aviser les personnes que vous avez côtoyées.\"},\"Body1\":\"Les personnes exposées n’obtiendront aucun renseignement sur vous \",\"Body2\":\"ni sur le moment de votre contact. Les personnes recevront uniquement une notification disant qu’elles ont récemment été exposées.\",\"CTA\":\"Suivant\",\"NoCode\":\"Vous n’avez pas de clé à usage unique?\"},\"IntermediateStep\":{\"Title\":\"Finir le partage d’expositions\",\"List\":{\"1a\":\"1. \",\"1b\":\"Fait.\",\"2a\":\"2. Vous fournissez quelques précisions \",\"2b\":\"qui aideront l’application à déterminer la période où vous étiez probablement le plus contagieux. Vous n’êtes pas obligé de le faire.\",\"3a\":\"3. Vous autorisez votre téléphone \",\"3b\":\"à partager vos codes aléatoires. Ceux-ci permettent d’aviser les personnes que vous avez côtoyées.\"},\"Body1\":\"Personne n’obtiendra de renseignements sur vous \",\"Body2\":\"ni sur le moment de votre contact. Les personnes recevront uniquement une notification disant qu’elles ont récemment été exposées.\",\"CTA\":\"Suivant\"},\"Step2\":{\"Title1\":\"Vous sentez-vous malade?\",\"Body1\":\"Symptômes possibles :\",\"Symptoms\":{\"1\":\"Toux ou difficultés respiratoires.\",\"2\":\"Fièvre ou frissons.\",\"3\":\"Fatigue ou faiblesse intense.\",\"4\":\"Courbatures ou mal de tête.\",\"5\":\"Perte soudaine de l’odorat ou du goût.\",\"6\":\"Maux d’estomac.\"},\"Option1\":\"Oui, vous avez eu 1 ou plusieurs symptômes aujourd’hui ou au cours des 14 derniers jours.\",\"Option2\":\"Non, vous n’avez eu aucun symptôme.\",\"Option3\":\"Vous n’êtes pas certain. \",\"Option4\":\"Vous préférez ne pas répondre.\",\"Title2\":\"Pourquoi cette question\",\"Body2\":\"Votre réponse permet de déterminer la période où vous étiez probablement le plus contagieux. L’application sait ainsi quelles expositions il faut partager pour n’aviser que les personnes côtoyées pendant cette période. Votre réponse ne sera jamais échangée ni conservée ailleurs que sur votre téléphone.\",\"CTA\":\"Suivant\"},\"FormView\":{\"Title\":\"Entrez votre clé à usage unique\",\"Body\":\"Entrez la clé reçue lors de votre diagnostic. Afin de prévenir les fausses notifications, vous ne recevrez une clé que si vous avez obtenu un résultat positif au test de dépistage de la COVID-19.\",\"Action\":\"Soumettre la clé\",\"InputLabel\":\"code Alerte COVID\"},\"OtkDiagnosedView\":{\"Title\":\"Vous avez déjà entré votre clé\",\"Body\":\"Il n’est pas nécessaire d’entrer la clé de nouveau.\",\"CTA\":\"Suivant\"},\"ConsentView\":{\"Title\":\"Vous pouvez maintenant téléverser vos codes\",\"Body1\":\"Votre téléphone vous demandera l’autorisation de téléverser vos codes ou « identifiants » aléatoires les plus récents.\",\"Body2a\":\"Personne ne recevra d’information sur vous \",\"Body2b\":\"ou sur le moment de votre proximité.\",\"Body3\":\"Vous pouvez aider à ralentir la propagation de la COVID-19 si vous acceptez.\",\"Action\":\"Accepter\"},\"TekUploadNoDate\":{\"Title\":\"Aviser les personnes côtoyées\",\"Body\":\"Vous n’avez pas fourni de date pour préciser la période où vous étiez probablement le plus contagieux.\\n\\nVous pouvez maintenant partager les expositions enregistrées par l’application pour aviser les personnes côtoyées au cours des 14 derniers jours.\",\"CTA\":\"Partager les expositions\"},\"SymptomOnsetDate\":{\"Title1\":\"Date de vos premiers symptômes\",\"Title2\":\"Pourquoi cette question\",\"Body1\":\"La date des premiers symptômes permet de déterminer la période où vous étiez probablement le plus contagieux. L’application sait ainsi quelles expositions il faut partager pour n’aviser que les personnes côtoyées pendant cette période.\\n\\nCette date ne sera jamais échangée ni conservée ailleurs que sur votre téléphone.\",\"CTA\":\"Partager les expositions\",\"CTA2\":\"Je préfère ne pas répondre\"},\"TestDate\":{\"Title1\":\"Date de votre test\",\"Body1\":\"Entrez la date de votre test de dépistage de la COVID-19.\",\"Title2\":\"Pourquoi cette question\",\"Body2\":\"La date de votre test permet de déterminer la période où vous étiez probablement le plus contagieux. L’application sait ainsi quelles expositions il faut partager pour n’aviser que les personnes côtoyées pendant cette période.\\n\\nCette date ne sera jamais échangée ni conservée ailleurs que sur votre téléphone.\",\"CTA\":\"Partager les expositions\",\"CTA2\":\"Je préfère ne pas répondre\"},\"NoCode\":{\"NoRegion\":{\"Title\":\"Vous n’avez pas choisi de province ou de territoire\",\"Body\":\"Dans certaines provinces et certains territoires, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nPour savoir si vous pouvez obtenir une clé à usage unique, vous devez sélectionner votre région.\",\"ChooseRegionCTA\":\"Sélectionner une province ou un territoire\"},\"RegionNotCovered\":{\"Title\":\"Pas encore de signalement dans votre région\",\"Body\":\"Les personnes de votre région ne peuvent pas encore obtenir de clé à usage unique.\\n\\nVous ne pourrez pas aviser les personnes d’une exposition. Mais vous pouvez tout de même vous isoler pendant 14 jours, et ainsi :\",\"Body2\":\"Assurer la sécurité de votre communauté.\",\"Body3\":\"Protéger les personnes et les animaux vivant avec vous.\",\"Body4\":\"Prendre soin de vous.\"}}},\"Notification\":{\"ExposedMessageTitle\":\"Vous avez été exposé\",\"ExposedMessageBody\":\"Vous avez eu un contact étroit avec une personne qui a signalé un diagnostic de COVID-19 au moyen de l’application. Informez-vous sur les étapes à suivre.\",\"OffMessageTitle\":\"Alerte COVID est désactivée\",\"OffMessageBody\":\"Activez Alerte COVID pour être avisé en cas d’exposition potentielle à la COVID-19.\",\"DailyUploadNotificationTitle\":\"Téléversez vos nouveaux codes aléatoires\",\"DailyUploadNotificationBody\":\"Aidez à ralentir la propagation de la COVID-19 en téléversant vos nouveaux codes aléatoires.\",\"ExposureChecksTitle\":\"Alerte COVID\",\"ExposureChecksBody\":\"Vérification d’exposition en cours\",\"ExposureChecksAndroidChannelName\":\"Vérifications d’exposition\",\"PausedMessageTitle\":\"Vous avez désactivé l’application\",\"PausedMessageBody\":\"L’application ne peut pas enregistrer d’expositions puisque vous l’avez désactivée.\",\"OTKNotSharedTitle\":\"Pas tout à fait fini\",\"OTKNotSharedBody\":\"Vous avez entré votre clé à usage unique, mais vous n’avez pas encore partagé vos expositions. Les personnes que vous avez côtoyées n’ont pas encore été notifiées.\",\"AndroidChannelName\":\"Alerte COVID\",\"OutbreakMessageTitle\":\"You have been exposed to an Outbreak\",\"OutbreakMessageBody\":\"There was an outbreak at one of the locations you scanned\"},\"Partners\":{\"Label\":\"Conçu en partenariat avec\"},\"BottomSheet\":{\"Collapse\":\"Fermer\",\"OnStatus\":\"Alerte COVID est ACTIVÉE, Appuyez pour voir le menu\",\"OffStatus\":\"Alerte COVID est DÉSACTIVÉE, Appuyez pour voir le menu\"},\"RegionLanding\":{\"Title\":\"Sélectionnez votre province ou territoire\",\"Body1\":\"Les gouvernements provinciaux et territoriaux s’efforcent de rendre Alerte COVID utilisable partout au Canada.\",\"Body2\":\"Dans certains endroits, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\",\"RegionSelectBtn\":\"Sélectionnez une province ou un territoire\"},\"RegionPicker\":{\"Title\":\"Où habitez-vous? (facultatif)\",\"SettingsTitle\":\"Province ou territoire\",\"Close\":\"Fermer\",\"Body\":\"Sélectionnez votre région pour savoir si vous pouvez signaler un diagnostic à l’aide de l’application.\",\"Skip\":\"Ignorer\",\"GetStarted\":\"Commencer\",\"Optional\":\"* Facultatif\",\"AB\":\"Alberta\",\"BC\":\"Colombie-Britannique\",\"MB\":\"Manitoba\",\"NB\":\"Nouveau-Brunswick\",\"NL\":\"Terre-Neuve-et-Labrador\",\"NT\":\"Territoires du Nord-Ouest\",\"NS\":\"Nouvelle-Écosse\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Île-du-Prince-Édouard\",\"QC\":\"Québec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\",\"None\":\"Ne rien sélectionner\",\"ExposedHelpLinkNote\":{\"AB\":\"(page en anglais)\",\"BC\":\"(page en anglais)\",\"SK\":\"(page en anglais)\"}},\"RegionPickerNoPT\":{\"Exposed\":{\"Title\":\"Trouvez les consignes de votre région\",\"Body\":\"Les mesures à prendre après une exposition dépendent de votre province ou de votre territoire.\"}},\"A11yList\":{\"Start\":\"Début de liste\",\"End\":\"Fin de liste\",\"Bullet\":\"Puce\"},\"MenuButton\":\"Menu\",\"QRCode\":{\"CTA\":\"Ajouter une visite\",\"Reader\":{\"Title\":\"Placez votre appareil photo devant le code QR\",\"Body\":\"\",\"Back\":\"Précédent\",\"Learn\":\"\"},\"CameraPermissions\":{\"Title\":\"Autoriser l’accès à l’appareil photo\",\"Body\":\"Pour scanner un lieu, l’application doit pouvoir utiliser l’appareil photo.\",\"Title2\":\"Vous hésitez à autoriser l’accès?\",\"Body2\":\"You can also use your phone’s camera app if you prefer not to let the app access your camera.\",\"CTA\":\"Suivant\",\"Cancel\":\"Pas maintenant\"},\"CameraPermissionDenied\":{\"Title\":\"Autorisation refusée\",\"Body\":\"Vous avez refusé qu’Alerte COVID utilise votre appareil photo.\\n\\nPour scanner un lieu, vous devez autoriser Alerte COVID à utiliser l’appareil photo dans les paramètres de votre téléphone.\\n\\nVous pouvez aussi utiliser directement votre appareil photo ou une application de lecture de codes QR.\",\"CTA\":\"Autoriser l’accès dans les paramètres\",\"CTA2\":\"À propos de l’ajout d’une visite\"},\"CameraPermissionAlert\":{\"Title\":\"« Alerte COVID » souhaite accéder à l’appareil photo\",\"Subtitle\":\"Alerte COVID doit accéder à l’appareil photo pour lire le code QR.\",\"Allow\":\"Autoriser\",\"DontAllow\":\"Ne pas autoriser\"},\"LearnAboutQRScan\":{\"Title\":\"Comment ajouter une visite\",\"Body1\":\"1. Autorisez Alerte COVID à accéder à l’appareil photo pour lire le code QR.\",\"Body2\":\"2. Placez votre appareil devant le code.\"},\"Error\":{\"Title\":\"Code QR non reconnu\",\"Body\":\"Vous n’avez peut-être pas scanné un code QR officiel d’Alerte COVID.\\n\\nDemandez à quelqu’un où se trouve l’affiche officielle d’Alerte COVID.\",\"CTA\":\"Réessayer\",\"CTA2\":\"Retour à l’écran d’accueil\"},\"CheckInView\":{\"Title\":\"Merci d’avoir ajouté une visite\",\"Body1\":\"Vous ne serez avisé que si quelqu’un présent pendant votre visite est déclaré positif prochainement. Vos visites \",\"Body2\":\"ne sont communiquées à personne.\\n\\n\",\"Body3\":\"N’oubliez pas de scanner à chaque visite.\",\"Body4\":\"Vous pouvez supprimer cette visite en tout temps en allant dans Vos visites.\",\"CTA1\":\"Retour à l’écran d’accueil\",\"CTA2\":\"Vos visites\",\"MoreInfo\":\"Annuler\"},\"Exposed\":{\"Title\":\"You've been Exposed\"},\"CancelCheckIn\":{\"Title\":\"FR Scan cancelled\"},\"ScanAPlace\":{\"Title\":\"Ajouter une visite\",\"Body\":\"Vous serez avisé si quelqu’un présent pendant votre visite est déclaré positif prochainement.\\n\\nL’application enregistrera vos visites sur votre téléphone, en plus de leur date et heure.\\n\\nVos visites **ne sont communiquées à personne.**\",\"CTA\":\"Suivant\",\"CTA2\":\"À propos de l’ajout d’une visite\"},\"ProximityExposure\":\"Passez un test\",\"OutbreakExposed\":{\"Title\":\"Exposition possible lors d’une visite\",\"Body\":\"Une personne déclarée positive était présente lors d’une de vos visites.\",\"CTA\":\"Résultat de test négatif?\",\"SelfIsolate\":{\"Title\":\"Isolez-vous et passez un test\",\"Body\":\"Vous êtes un contact à risque élevé. Si vous avez des symptômes, faites-vous tester immédiatement. Si vous n’en avez pas, attendez 7 jours après la date d’exposition pour vous faire tester.\\n\\nVous devez vous isoler pendant 14 jours à partir de la date d’exposition.\\n\\nPour toute question, contactez le bureau de santé de votre région.\"},\"GetTested\":{\"Title\":\"Passez un test\",\"Body\":\"Vous devez passer un test de dépistage dès que possible.\\n\\nEntre-temps, évitez les contacts avec les personnes vulnérables.\\n\\nPour toute question, contactez le bureau de santé de votre région.\"},\"SelfMonitor\":{\"Title\":\"Évitez les contacts et surveillez vos symptômes\",\"Body\":\"Vous devez éviter les contacts avec les personnes vulnérables et surveiller vos symptômes. Si vous en développez, remplissez une autoévaluation.\\n\\nPour toute question, contactez le bureau de santé de votre région.\"}}},\"PlacesLog\":{\"Title\":\"Vos visites\",\"Body1\":\"Vous serez avisé s’il y a eu un risque d’exposition lors des visites qui suivent.\",\"Body2a\":\"Vos visites ne sont communiquées \",\"Body2b\":\"à personne. Elles sont uniquement conservées sur votre téléphone.\",\"DeleteBtnCTA\":\"Supprimer toutes les visites\",\"DeleteIcon\":\"Supprimer\",\"NoVisit\":\"Aucune visite\",\"Back\":\"Précédent\",\"Alert\":{\"Title\":\"Voulez-vous vraiment supprimer cette visite?\",\"TitleDeleteAll\":\"Voulez-vous vraiment supprimer toutes les visites?\",\"Subtitle\":\"S’il y a un risque d’exposition dans ce lieu, vous n’en serez pas avisé.\",\"SubtitleDeleteAll\":\"S’il y a un risque d’exposition dans un de ces lieux, vous n’en serez pas avisé.\",\"Cancel\":\"Annuler\",\"ConfirmDelete\":\"Oui, supprimer\",\"ConfirmDeleteAll\":\"Oui, tout supprimer\"}},\"ExposureHistory\":{\"Title\":\"Vos expositions récentes\",\"Body\":\"Voici la liste de vos expositions pour les 14 derniers jours.\",\"NoExposures\":\"Aucune exposition\"},\"Errors\":{\"Action\":\"OK\",\"TekUploadNoDate\":{\"Title\":\"\",\"Body\":\"Vous devez sélectionner une date pour pouvoir partager les expositions.\"},\"TekUploadOffline\":{\"Title\":\"Problème de connexion\",\"Body\":\"Vos codes aléatoires n’ont pas été téléversés. Vérifiez que votre téléphone est connecté à Internet au moyen du Wi-Fi ou de vos données.\"},\"TekUploadServer\":{\"Title\":\"Réessayez plus tard\",\"Body\":\"Il y a eu un problème avec le système.\"},\"TekUploadPermission\":{\"Title\":\"Les codes aléatoires n’ont pas pu être téléversés\",\"Body\":\"Vous n’avez pas donné votre autorisation.\"},\"OtcCodeNotEntered\":{\"Title\":\"\",\"Body\":\"Vous devez entrer votre clé pour pouvoir la soumettre.\"},\"OtcUploadInvalidOneTimeCode\":{\"Title\":\"Clé à usage unique non reconnue\",\"Body\":\"Votre clé n’a pas pu être reconnue. Veuillez essayer à nouveau.\"},\"OtcUploadTemporaryBan\":{\"Title\":\"Attendez pendant 1 heure\",\"Body\":\"Vous avez essayé d’entrer une clé à usage unique trop de fois.\"},\"OtcUploadDefault\":{\"Title\":\"Réessayez plus tard\",\"Body\":\"Il y a eu un problème avec le système.\"},\"OtcUploadOffline\":{\"Title\":\"Problème de connexion\",\"Body\":\"Votre clé n’a pas été envoyée. Vérifiez que votre téléphone est connecté à Internet au moyen du Wi-Fi ou de vos données.\"},\"ContentNoConnect\":{\"Title\":\"Impossible d’accéder aux informations de votre province ou de votre territoire\",\"Body\":\"Réessayez d’y accéder plus tard.\"},\"DataSharingStep2\":{\"Title\":\"\",\"Body\":\"Vous devez sélectionner une réponse pour passer à l’étape suivante.\"},\"ApiNotConnected\":{\"Title\":\"Une erreur s'est produite\",\"Body\":\"Malheureusement, il semble qu'Alerte COVID ne fonctionne pas sur votre téléphone. \\n\\nSi vous avez plusieurs profils utilisateur, essayez d'installer l'application dans le profil de base.\"}}}}")
+export default JSON.parse("{\"en\":{\"Home\":{\"AppName\":\"COVID Alert\",\"UserStopped\":{\"Title\":\"You turned off the app\",\"CTA\":\"Turn back on\"},\"NoExposureDetected\":{\"AllSetTitle\":\"You’re all set\",\"RegionCovered\":{\"Title\":\"No exposure detected\",\"Body\":\"You have not been near anyone who reported a COVID-19 diagnosis through this app.\",\"GuidanceUrl\":\"https://www.canada.ca/en/public-health/services/publications/diseases-conditions/covid-19-how-to-isolate-at-home.html\",\"AllSetBody\":\"Your phone is now running Bluetooth to collect codes from nearby phones.\"},\"NoRegion\":{\"Title\":\"No exposure detected\",\"Body\":\"You have not been near anyone who reported a COVID-19 diagnosis through this app.\\n\\nHowever, in some provinces and territories, people cannot yet report COVID-19 diagnoses through this app.\",\"AllSetBody\":\"We’ll let you know as soon as people outside Ontario can report a diagnosis through COVID Alert.\\n\\nIt’s still helpful to keep COVID Alert on.\\n\\nThat way, when people are able to report a diagnosis, you’ll find out if you were near them.\"},\"RegionNotCovered\":{\"Title\":\"No reporting yet in your area\",\"Body\":\"People in your province or territory are not yet able to report a COVID-19 diagnosis through this app.\\n\\nIt’s still helpful to turn COVID Alert on. That way, when people are able to report a diagnosis, you’ll be notified if you were near them.\",\"GuidanceUrl\":\"https://www.canada.ca/en/public-health/services/publications/diseases-conditions/covid-19-how-to-isolate-at-home.html\",\"AllSetBody\":\"We’ll let you know as soon as people in your area can report a diagnosis through COVID Alert.\\n\\nIt’s still helpful to keep COVID Alert on.\\n\\nThat way, when people are able to report a diagnosis, you’ll find out if you were near them.\"}},\"ExposureDetected\":{\"Title\":\"You’ve been exposed\",\"Body1\":\"You’ve been near at least one person who reported a COVID-19 diagnosis through the app. You were near them for at least 15 minutes.\",\"Title2\":\"What next?\",\"Body2\":\"You’re at risk of being infected.\",\"Notification\":{\"Received\":\"Notification received\"},\"NegativeTest\":{\"CTA\":\"Negative test?\"},\"Dismiss\":{\"Title\":\"Negative test result?\",\"Body\":\"Dismiss this alert to receive any new exposure notifications. Dismissing this alert will not delete the exposure logs on your phone.\",\"Body2\":\"You can also dismiss the alert if you’re following your area’s guidance.\",\"CTA\":\"Dismiss alert\",\"Confirm\":{\"Body\":\"Are you sure you want to dismiss this alert?\",\"Accept\":\"Yes, dismiss\",\"Cancel\":\"Cancel\"}}},\"ChooseRegionCTA\":\"Choose province or territory\",\"How\":\"Diagnosed with COVID-19?\",\"CTA\":\"What to do next\",\"HowUrl\":\"https://www.canada.ca/\",\"ExposedHelpCTA\":\"What to do if you’re exposed\",\"DiagnosedView\":{\"Title\":\"Thanks for helping fight COVID-19\",\"Body1\":{\"One\":\"Tomorrow, you’ll get your last notification asking permission to upload any new random codes your phone sent out.\",\"Other\":\"For the next {number} days, you’ll get a daily notification asking permission to share any new exposures your phone recorded.\"},\"Body2\":\"You do not have to agree, even if you agreed the day before.\",\"Body3\":\"If you’re worried about your symptoms, get medical care as soon as you can. Take care of yourself and your family during this time.\",\"Tip\":{\"Title\":\"TIP: \"}},\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DiagnosedShareView\":{\"Title\":\"Share new exposures\",\"Body1\":\"Thanks for helping slow the spread!\",\"Body2\":\"Reminder: \",\"Body3\":\"To help keep people safe, you need to share your new exposures every day, especially if you have to leave your home.\",\"ButtonCTA\":\"Share new exposures\"},\"DiagnosedShareUploadView\":{\"Title\":\"Not quite done\",\"Body1\":\"You entered your one-time key but you have not shared any exposures yet. People who were near you have not been notified yet.\",\"ButtonCTA\":\"Finish sharing exposures\"},\"BluetoothDisabled\":\"Bluetooth is off\",\"EnableBluetoothCTA\":\"You need to turn on Bluetooth in your phone’s settings so COVID Alert can work.\",\"TurnOnBluetooth\":\"Turn on Bluetooth\",\"EnUnauthorized\":{\"Title\":\"COVID Alert is off\",\"Body1\":\"From the button below, choose Exposure Notifications, then scroll down and turn on \\\"Share Exposure Information\\\".\",\"CTA\":\"Enable COVID Alert\"},\"EnDisabled\":{\"Title\":\"COVID Alert is off\",\"Body1\":\"You did not enable COVID Alert. The app needs your permission to work.\",\"CTA\":\"Enable COVID Alert\",\"AndroidTitle2\":\"Not sure about turning on Location?\",\"AndroidBody1\":\"You need to have Location turned on for COVID Alert to work. To use Bluetooth scanning, Android phones need Location setting on for all apps.\",\"AndroidBody2a\":\"Despite that, COVID Alert has \",\"AndroidBody2b\":\"no way of knowing where you are. \",\"AndroidBody2c\":\"You can check the app’s permissions in your phone’s settings. You’ll see it does not have permission for location services.\"},\"FrameworkUnavailable\":{\"Title\":\"Something went wrong\",\"Body\":\"To use COVID Alert, you need to update your phone’s operating system.\"},\"UnknownProblem\":{\"Title\":\"Something went wrong\",\"Body\":\"It looks like your phone cannot support COVID Alert.\",\"CTA\":\"Get Help\"},\"NoConnectivity\":\"COVID Alert is offline\",\"NoConnectivityDetailed\":\"COVID Alert only checks for exposure if your phone is connected to the Internet.\",\"ExternalLinkHint\":\"Opens in a new window\",\"LastCheckedMinutes\":{\"One\":\"Last checked {number} minute ago\",\"Other\":\"Last checked {number} minutes ago\"},\"LastCheckedHours\":{\"One\":\"Last checked {number} hour ago\",\"Other\":\"Last checked {number} hours ago\"},\"LastCheckedDays\":{\"One\":\"Last checked {number} day ago\",\"Other\":\"Last checked {number} days ago\"}},\"OverlayClosed\":{\"SystemStatus\":\"COVID Alert is \",\"SystemStatusOn\":\"active\",\"SystemStatusOff\":\"off\",\"NotificationStatus\":\"Notifications are \",\"NotificationStatusOff\":\"off\",\"TapPrompt\":\"Tap for menu\"},\"OverlayOpen\":{\"Version\":\"Version\",\"BluetoothCardBody\":\"You need to turn on Bluetooth in your phone’s settings so COVID Alert can work.\",\"BluetoothCardAction\":\"Turn on Bluetooth\",\"EnterCodeCardTitle\":\"Diagnosed with COVID-19?\",\"EnterCodeCardBody\":\"To let people know they’ve been exposed, you need a one-time key. We’ll also ask for details to help narrow down when you were likely most contagious.\",\"EnterCodeCardAction\":\"Enter your one-time key\",\"EnterCodeCardTitleDiagnosed\":\"You’re helping stop COVID-19\",\"EnterCodeCardBodyDiagnosed\":\"Thank you for doing your part to slow the spread. \",\"EnterCodeCardDiagnosedCountdown\":{\"One\":\"{number} day left!\",\"Other\":\"{number} days left!\"},\"CodeNotShared\":{\"Title\":\"Help slow the spread\",\"Body\":\"You entered your one-time key but you have not shared any exposures yet.\",\"CTA\":\"Finish sharing exposures\"},\"NotificationCardStatus\":\"Notifications are off\",\"NotificationCardStatusOff\":\"OFF\",\"NotificationCardBody\":\"You will not receive notifications if you’ve been near anyone who has reported a COVID-19 diagnosis. Finding out potential exposures right away is essential to slowing the spread of COVID-19.\",\"NotificationCardAction\":\"Turn on notifications\",\"ExposureNotificationCardStatus\":\"Exposure notifications are \",\"ExposureNotificationCardBody\":\"COVID Alert cannot notify you as soon as it learns of a possible exposure. Finding out right away is essential to slowing the spread of COVID-19.\",\"ExposureNotificationCardAction\":\"Turn on exposure notifications\",\"NoConnectivityCardAction\":\"COVID Alert is offline\",\"NoConnectivityCardBody\":\"Connect to the Internet to notify people or check for exposures.\",\"EnUnauthorizedCardAction\":\"Enable COVID Alert\",\"EnUnauthorizedCardBody\":\"Choose Exposure Notifications, then scroll down and turn on \\\"Share Exposure Information\\\".\",\"ShareDiagnosisCode\":{\"CtaReport\":\"Enter your one-time key\",\"CtaFinish\":\"Finish sharing exposures\"}},\"Info\":{\"CheckSymptoms\":\"Check symptoms\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/en\",\"TellAFriend\":\"Share app\",\"Support\":\"Support\",\"LearnMore\":\"How it works\",\"Help\":\"Help\",\"HelpUrl\":\"https://www.canada.ca/en/public-health/services/diseases/coronavirus-disease-covid-19/covid-alert/help.html\",\"GetCode\":\"Get a one-time key\",\"YourVisits\":\"Your Visits\",\"YourRecentExposures\":\"Your recent exposures\",\"ChangeLanguage\":\"Change language\",\"ToggleCovidAlert\":{\"TurnOff\":\"Turn off COVID Alert\",\"TurnOn\":\"Turn on COVID Alert\",\"Confirm\":{\"Title\":\"Turn off COVID Alert?\",\"Body\":\"Make sure you’re wearing full medical PPE (personal protective equipment) or you’re only with your household.  The app will not be able to record exposures.\",\"Cancel\":\"Keep it on\",\"Accept\":\"Turn it off\"}},\"ChangeRegion\":\"Change province or territory\",\"Privacy\":\"Privacy\",\"PrivacyUrl\":\"https://www.canada.ca/en/public-health/services/diseases/coronavirus-disease-covid-19/covid-alert/privacy-policy.html\",\"SettingsTitle\":\"Settings\",\"InformationTitle\":\"About COVID Alert\"},\"Landing\":{\"En\":\"English\",\"Fr\":\"Français\",\"AltText\":\"Illustration of 8 people enjoying a public space. There’s a person with a walking cane, a person in a wheelchair, a person wearing a hijab, several people of colour, and a representation of gender diversity.\",\"CanadaAltText\":\"Symbol of the Government of Canada\"},\"LanguageSelect\":{\"Title\":\"Language\",\"En\":\"English (Canada)\",\"Fr\":\"Français (Canada)\",\"EnShort\":\"English\",\"FrShort\":\"Français\",\"Close\":\"Close\"},\"Tutorial\":{\"step-1Title\":\"How COVID Alert works\",\"step-1AltText\":\"A hand-drawn illustration of a mobile phone in a pocket, sending and receiving signals.\",\"step-1\":\"The app runs in the background and will not interrupt your activities.\\n\\nWhenever you’re near someone else with COVID Alert, both phones exchange random codes every 5 minutes.\\n\\nThe random codes change often and cannot be used to identify you. \",\"step-2Title\":\"What’s an exposure?\",\"step-2AltText\":\"A hand-drawn illustration of 2 mobile phones side-by-side with a timer clock connecting both phones.\",\"step-2\":\"The app estimates how near people are by the strength of Bluetooth signals.\\n\\nThe app will record an exposure if the contact:\\n• Is closer than 2 metres.\\n• Lasts at least 15 minutes.\\n• Happens during the time the person was likely most contagious.\",\"step-3Title\":\"Getting a positive test\",\"step-3AltText\":\"A hand-drawn illustration of a mobile phone uploading a numeric code to a cloud.\",\"step-3\":\"If a person with the app is diagnosed with COVID-19, they can choose to upload the random codes their phone sent.\\n\\nThey can also give details to help narrow down when they were likely most contagious. Those details tell the phone which random codes to upload.\\n\\nThe random codes go into a central server. No other information is sent to the server.\",\"step-4Title\":\"Looking for exposures\",\"step-4AltText\":\"A hand-drawn illustration of a mobile phone uploading a numeric code to a cloud.\",\"step-4\":\"Every day, whenever it has an Internet connection, your phone will get a list of the random codes from people who reported a diagnosis.\\n\\nIf it finds codes that match, the app notifies you that you’ve been exposed and explains what to do next.\",\"ActionBack\":\"Back\",\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"Close\":\"Close\"},\"QRCodeOnboarding\":{\"step-1Title\":\"New pilot in Mooseville only: exposures from public places\",\"step-1AltText\":\"-\",\"step-1\":\"The app can notify you if you’ve been at a place at the same time as an exposure.\\n\\nThis kind of exposure notification uses QR codes you scan as you go about your day.\\n\\nIt’s more private than sharing your contact details with the places you visit.\",\"step-2Title\":\"How QR code notifications work\",\"step-2AltText\":\"-\",\"step-2\":\"Public health identifies when a place has an exposure risk.\\n\\nEvery day, your app downloads a list of those places and times.\\n\\nIf there’s a match with one of your visits, the app notifies you and explains what to do next.\",\"step-3Title\":\"Your privacy is protected\",\"step-3AltText\":\"-\",\"step-3\":\"The app saves your visits on your phone, including the dates and times you were there.\\n\\n**Your visits and exposures are not shared** with anyone.\\n\\nYou can get to your recent visits from the menu. If you need to, you can delete any visit.\",\"step-4Title\":\"Help make contact tracing easier\",\"step-4AltText\":\"-\",\"step-4\":\"Next time you’re out and about, look for COVID Alert posters with QR codes.\\n\\nStart adding your visits and get notified of possible exposures from them.\",\"step-4-banner\":\"Right now, posters with QR codes are only available in _____ as part of a small-scale pilot project.\",\"ActionBack\":\"Back\",\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"Close\":\"Close\"},\"Onboarding\":{\"Step\":\"Step\",\"Of\":\"of\",\"ActionNext\":\"Next\",\"ActionEndSkip\":\"Skip\",\"ActionEnd\":\"Done\",\"ActionBack\":\"Back\",\"Start\":{\"Title\":\"Together, let’s stop the spread of COVID-19\",\"ImageAltText\":\"A hand-drawn illustration with lots of different people in it. Dotted lines show their paths crossing and leading to a smartphone below.\",\"Body1\":\"COVID Alert helps us break the cycle of infection. The app can let people know of possible exposures before any symptoms appear. \",\"Body2\":\"That way, we can take care of ourselves and protect our communities.\"},\"WhatItsNot\":{\"Title\":\"What COVID Alert does not do\",\"ImageAltText\":\"A hand-drawn illustration with many houses and a high-rise building on a curving street. Small message bubbles appear above each home.\",\"Body1\":\"The app will not tell you in the moment if you’re currently near someone who’s been diagnosed.\",\"Body2\":\"It will not tell you about outbreaks in your community.\"},\"Anonymous\":{\"Title\":\"Your privacy is protected\",\"ImageAltText\":\"A hand-drawn illustration of shaded glasses and hat to go incognito. Above the hat, a location pin icon is crossed out.\",\"Body1\":\"COVID Alert **does not use GPS** or track your location.\",\"Body2\":\"It has **no way of knowing**:\",\"Bullet1\":\"Your location.\",\"Bullet2\":\"Your name or address.\",\"Bullet3\":\"Your phone’s contacts.\",\"Bullet4\":\"Your health information, except what you decide to share.\",\"Bullet5\":\"The health information of anyone you’re near.\"},\"HowItWorks\":{\"Title\":\"How it works\",\"ImageAltText\":\"A hand-drawn illustration of 3 mobile phones sending out random codes via bluetooth.\",\"Body1\":\"The app uses Bluetooth to exchange random codes with nearby phones.\",\"Body2\":\"Every day, it checks a list of random codes from people who tell the app they tested positive.\",\"Body3\":\"If you’ve been near one of those people within the last 14 days, you’ll get a notification.\",\"HowItWorksCTA\":\"Learn more about how it works\"},\"PartOf\":{\"Title\":\"One part of public health\",\"ImageAltText\":\"A hand-drawn illustration of different ways of fighting COVID-19, including hand sanitizer, handwashing, face mask and the COVID Alert app.\",\"Body1\":\"COVID Alert is just one part of the public health effort to stop the spread of COVID-19.\",\"Body2\":\"Follow all public health guidelines in your area.\",\"Body3\":\"COVID Alert does not replace medical advice. If you get sick, contact your doctor or other healthcare provider.\"},\"Permissions\":{\"Title\":\"The app will now ask your permission\",\"ImageAltText\":\"A hand-drawn illustration of a mobile phone displaying a dialogue box with a red X and a green check mark on the screen.\",\"Body1\":\"**Allow the app** to start logging random codes or “random IDs” when you’re near other phones. The app will access the date, duration and signal strength related to the random codes, but they never leave your phone.\",\"Body2\":\"You’ll also need to **let the app send you notifications**.\",\"PrivacyButtonCTA\":\"Learn more about privacy\"}},\"Privacy\":{\"Title\":\"Privacy policy\",\"Close\":\"Close\"},\"ThankYou\":{\"Title\":\"Thank you for helping\",\"Body\":\"Exposure notifications are on. You will receive a message if COVID Alert detects that you’ve been near someone who has reported a COVID-19 diagnosis.\",\"Dismiss\":\"Dismiss\"},\"DataUpload\":{\"Cancel\":\"Cancel\",\"Close\":\"Close\",\"Back\":\"Back\",\"InfoSmall\":\"Your random codes will not be uploaded unless you give permission in the next step.\",\"ShareToast\":\"Random codes uploaded successfully\",\"StepXofY\":\"Step {x} of {y}\",\"Today\":\"Today\",\"Yesterday\":\"Yesterday\",\"Earlier\":\"Even earlier\",\"Step0\":{\"Title\":\"Notify people they’ve been exposed\",\"List\":{\"1a\":\"1. Enter the one-time key \",\"1b\":\"you got when you were diagnosed.\",\"2a\":\"2. Give the app details \",\"2b\":\"to help narrow down when you were likely most contagious. You can choose not to give any details.\",\"3a\":\"3. Allow your phone to share \",\"3b\":\"your random codes. These codes notify people you’ve been near.\"},\"Body1\":\"Exposed people will not get any information about you \",\"Body2\":\"or the time you were near them. They will only get a notification that they were exposed recently.\",\"CTA\":\"Next\",\"NoCode\":\"Need a one-time key?\"},\"IntermediateStep\":{\"Title\":\"Finish sharing exposures\",\"List\":{\"1a\":\"1. \",\"1b\":\"Done.\",\"2a\":\"2. Give the app details \",\"2b\":\"to help narrow down when you were likely most contagious. You can choose not to give any details.\",\"3a\":\"3. Allow your phone to share \",\"3b\":\"your random codes. These codes notify people you’ve been near.\"},\"Body1\":\"Nobody will get any information about you \",\"Body2\":\"or the time you were near them. They will only get a notification that they were exposed recently.\",\"CTA\":\"Next\"},\"Step2\":{\"Title1\":\"Are you feeling sick?\",\"Body1\":\"Symptoms could include:\",\"Symptoms\":{\"1\":\"Cough or breathing problems.\",\"2\":\"Fever or chills.\",\"3\":\"Feeling really tired or weak.\",\"4\":\"Sore muscles or headache.\",\"5\":\"New loss of smell or taste.\",\"6\":\"Stomach problems.\"},\"Option1\":\"Yes, you have or had at least 1 symptom within the last 14 days.\",\"Option2\":\"No, you’ve not had any symptoms.\",\"Option3\":\"You're not sure.\",\"Option4\":\"You prefer not to say.\",\"Title2\":\"Why we’re asking\",\"Body2\":\"This information helps figure out when you were likely most contagious. It tells the app which exposures to share so only people who were near you during that time will be notified. It will never be sent or stored outside your phone.\",\"CTA\":\"Next\"},\"FormView\":{\"Title\":\"Enter your one-time key\",\"Body\":\"Enter the key you got when you were diagnosed. To prevent false notifications, you can only get a key if you test positive for COVID-19.\",\"Action\":\"Submit key\",\"InputLabel\":\"COVID Alert key\"},\"OtkDiagnosedView\":{\"Title\":\"You've already entered the key\",\"Body\":\"You do not need to enter it again.\",\"CTA\":\"Next\"},\"ConsentView\":{\"Title\":\"Now you can upload your codes\",\"Body1\":\"Your phone will ask your permission to upload your most recent random codes or “random IDs”.\",\"Body2a\":\"Nobody will get any information about you \",\"Body2b\":\"or the time you were near them.\",\"Body3\":\"You can help slow the spread of COVID-19 if you agree.\",\"Action\":\"Agree\"},\"TekUploadNoDate\":{\"Title\":\"Notify people you've been near\",\"Body\":\"You did not give a date to help narrow down when you were likely most contagious.\\n\\nShare your exposures now to notify the people you were near in the last 14 days.\",\"CTA\":\"Share exposures\"},\"SymptomOnsetDate\":{\"Title1\":\"Date your symptoms started\",\"Title2\":\"Why we’re asking\",\"Body1\":\"This date helps figure out when you were likely most contagious. It tells the app which exposures to share so only people who were near you during that time will be notified.\\n\\nThe date does not get sent or stored anywhere outside your phone.\",\"CTA\":\"Share exposures\",\"CTA2\":\"Prefer not to say\"},\"TestDate\":{\"Title1\":\"Your test date\",\"Body1\":\"Enter the date you got your COVID-19 test.\",\"Title2\":\"Why we’re asking\",\"Body2\":\"Your test date helps figure out when you were likely most contagious. It tells the app which exposures to share so only people who were near you then will be notified.\\n\\nThe date does not get sent or stored anywhere outside your phone.\",\"CTA\":\"Share exposures\",\"CTA2\":\"Prefer not to say\"},\"NoCode\":{\"NoRegion\":{\"Title\":\"You skipped choosing province or territory\",\"Body\":\"In some provinces and territories, people cannot yet report a COVID-19 diagnosis through this app.\\n\\nTo find out if you can get a one-time key, you need to choose your area.\",\"ChooseRegionCTA\":\"Choose province or territory\"},\"RegionNotCovered\":{\"Title\":\"No reporting yet in your area\",\"Body\":\"People in your area cannot yet get a one-time key.\\n\\nYou will not be able to let people know they were exposed. But you can still self-isolate for 14 days, and:\",\"Body2\":\"Keep your community safe.\",\"Body3\":\"Protect people and pets you live with.\",\"Body4\":\"Take care of yourself.\"}}},\"Notification\":{\"ExposedMessageTitle\":\"You’ve been exposed\",\"ExposedMessageBody\":\"You’ve had close contact with someone who reported a COVID-19 diagnosis through the app. Learn more about what to do next.\",\"OffMessageTitle\":\"COVID Alert is off\",\"OffMessageBody\":\"Turn on COVID Alert to get notified if you’ve been near someone who has reported a COVID-19 diagnosis.\",\"DailyUploadNotificationTitle\":\"Upload new random codes\",\"DailyUploadNotificationBody\":\"Help slow the spread of COVID-19 and upload your new random codes.\",\"ExposureChecksTitle\":\"COVID Alert\",\"ExposureChecksBody\":\"Checking for exposures\",\"ExposureChecksAndroidChannelName\":\"Checking for exposures\",\"PausedMessageTitle\":\"You turned off the app\",\"PausedMessageBody\":\"The app is not recording exposures because you turned it off.\",\"OTKNotSharedTitle\":\"Not quite done\",\"OTKNotSharedBody\":\"You entered your one-time key but you have not shared any exposures yet. People who were near you have not been notified yet.\",\"AndroidChannelName\":\"COVID Alert\",\"OutbreakMessageTitle\":\"You have been exposed to an Outbreak\",\"OutbreakMessageBody\":\"There was an outbreak at one of the locations you scanned\"},\"Partners\":{\"Label\":\"Developed in partnership with\"},\"BottomSheet\":{\"Collapse\":\"Close\",\"OnStatus\":\"COVID Alert is on, Tap for more information\",\"OffStatus\":\"COVID Alert is off, Tap for more information\"},\"RegionLanding\":{\"Title\":\"Choose your province or territory\",\"Body1\":\"Provincial and territorial governments are working to support COVID Alert across Canada.\",\"Body2\":\"In some places, people cannot yet report a COVID-19 diagnosis through this app.\",\"RegionSelectBtn\":\"Choose province or territory\"},\"RegionPicker\":{\"Title\":\"Where do you live? (optional)\",\"SettingsTitle\":\"Province or territory\",\"Close\":\"Close\",\"Body\":\"Choose your region to check if you can report a diagnosis through this app.\",\"Skip\":\"Skip\",\"GetStarted\":\"Get started\",\"Optional\":\"* Optional\",\"AB\":\"Alberta\",\"BC\":\"British Columbia\",\"MB\":\"Manitoba\",\"NB\":\"New Brunswick\",\"NL\":\"Newfoundland and Labrador\",\"NT\":\"Northwest Territories\",\"NS\":\"Nova Scotia\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Prince Edward Island\",\"QC\":\"Quebec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\",\"None\":\"Prefer not to say\",\"ExposedHelpLinkNote\":{\"AB\":\"\",\"BC\":\"\",\"SK\":\"\"}},\"RegionPickerNoPT\":{\"Exposed\":{\"Title\":\"Find your area’s advice\",\"Body\":\"The steps you need to take when you’re exposed depend on the province or territory you live in.\"}},\"A11yList\":{\"Start\":\"List start\",\"End\":\"List end\",\"Bullet\":\"Bullet\"},\"MenuButton\":\"Menu\",\"QRCode\":{\"CTA\":\"Add a visit\",\"Reader\":{\"Title\":\"Point the camera at the QR code\",\"Body\":\"Subscribe to a place to receive notification if other people visit this place at a similar time as you.\",\"Back\":\"Back\",\"Learn\":\"Learn more about QR codes and COVID Alert\"},\"CameraPermissions\":{\"Title\":\"Give access to camera\",\"Body\":\"To scan a place, the app needs to use your camera.\",\"Title2\":\"Not sure about giving access?\",\"Body2\":\"You can also use your phone’s camera directly if you prefer not to let the app access your camera.\",\"CTA\":\"Next\",\"Cancel\":\"Not now\"},\"CameraPermissionDenied\":{\"Title\":\"Access denied\",\"Body\":\"You said no to COVID Alert using your phone’s camera.\\n\\nTo scan a place, you’ll need to update your phone’s settings to let COVID Alert  use the camera.\\n\\nYou can also use your phone’s camera directly or a QR scanning app.\",\"CTA\":\"Allow camera access in settings\",\"CTA2\":\"About adding a visit\"},\"CameraPermissionAlert\":{\"Title\":\"« COVID Alert » Would Like to Access the Camera\",\"Subtitle\":\"COVID Alert needs access to your camera so it can read the QR code.\",\"Allow\":\"Allow\",\"DontAllow\":\"Don’t Allow\"},\"LearnAboutQRScan\":{\"Title\":\"How to add a visit\",\"Body1\":\"1. Give COVID Alert access to your camera so it can read the QR code.\",\"Body2\":\"2. Point your phone at the code.\"},\"Error\":{\"Title\":\"QR code not recognized\",\"Body\":\"You may not have scanned an official COVID Alert QR code.\\n\\nAsk someone for the official COVID Alert poster.\",\"CTA\":\"Try Again\",\"CTA2\":\"Back to Home\"},\"CheckInView\":{\"Title\":\"Thank you for adding a visit\",\"Body1\":\"You’ll only be notified if public health finds an exposure risk from your visit. \",\"Body2\":\"Your visits are not shared \",\"Body3\":\"with anyone.\\n\\nRemember to scan each time you visit.\",\"Body4\":\"You can delete this visit any time by going into Your visits.\",\"CTA1\":\"Back to home\",\"CTA2\":\"Your visits\",\"MoreInfo\":\"Learn more about QR codes and COVID alert\"},\"Exposed\":{\"Title\":\"You've been Exposed\"},\"CancelCheckIn\":{\"Title\":\"Scan cancelled\"},\"ScanAPlace\":{\"Title\":\"Add a visit: Mooseville only\",\"Body\":\"Get notified if someone from today tests positive later.\\n\\nThe app saves your visits on your phone, including the dates and times you were there.\\n\\n**Your visits are not shared** with anyone.\",\"Banner\":\"Right now, posters with QR codes are only available in Mooseville as part of a small-scale pilot project.\",\"CTA\":\"Next\",\"CTA2\":\"About adding a visit\"},\"ProximityExposure\":\"Get tested\",\"OutbreakExposed\":{\"Title\":\"Possible exposure at a visit\",\"Body\":\"Someone tested positive from one of your visits.\",\"CTA\":\"Negative test?\",\"SelfIsolate\":{\"Title\":\"Isolate and get tested\",\"Body\":\"You’re a high-risk contact. If you have symptoms, get tested right away. If you do not have symptoms, wait to get tested at least 7 days after the date of exposure.\\n\\nYou also need to isolate for 14 days from the date of exposure.\\n\\nIf you have questions, contact your local public health unit.\"},\"GetTested\":{\"Title\":\"Get tested\",\"Body\":\"You need to get tested as soon as possible.\\n\\nIn the meantime, avoid contact with vulnerable people.\\n\\nIf you have questions, contact your local public health unit.\"},\"SelfMonitor\":{\"Title\":\"Avoid people and self-monitor\",\"Body\":\"You need to avoid contact with vulnerable people and watch yourself carefully for symptoms. If you get symptoms, take a self-assessment.\\n\\nIf you have questions, contact your local public health unit.\"}}},\"PlacesLog\":{\"Title\":\"Your Visits\",\"Body1\":\"You’ll get notified if there’s a possible exposure from the following visits.\",\"Body2a\":\"Your visits are not shared \",\"Body2b\":\"with anyone. They’re only on your phone.\",\"DeleteBtnCTA\":\"Delete all visits\",\"DeleteIcon\":\"Delete\",\"NoVisit\":\"No visits yet\",\"Back\":\"Back\",\"Alert\":{\"Title\":\"Are you sure you want to delete this visit?\",\"TitleDeleteAll\":\"Are you sure you want to delete all visits?\",\"Subtitle\":\"If there’s a possible exposure at this place, you will not be notified.\",\"SubtitleDeleteAll\":\"If there’s a possible exposure at any of these places, you will not be notified.\",\"Cancel\":\"Cancel\",\"ConfirmDelete\":\"Yes, delete\",\"ConfirmDeleteAll\":\"Yes, delete all\"}},\"ExposureHistory\":{\"Title\":\"Your recent exposures\",\"Body\":\"This list shows your exposures from the last 14 days.\",\"NoExposures\":\"No exposures\"},\"Errors\":{\"Action\":\"OK\",\"TekUploadNoDate\":{\"Title\":\"\",\"Body\":\"You need to choose a date before you can share exposures.\"},\"TekUploadOffline\":{\"Title\":\"Lost connection\",\"Body\":\"Your random codes were not uploaded. Make sure your phone is connected to the Internet, through either data or WiFi.\"},\"TekUploadServer\":{\"Title\":\"Try again later\",\"Body\":\"Something went wrong with the system.\"},\"TekUploadPermission\":{\"Title\":\"Random codes could not be uploaded\",\"Body\":\"You did not give permission.\"},\"OtcCodeNotEntered\":{\"Title\":\"\",\"Body\":\"You need to enter your key before you can submit.\"},\"OtcUploadInvalidOneTimeCode\":{\"Title\":\"Incorrect key\",\"Body\":\"Your one-time key could not be recognized. Please try again.\"},\"OtcUploadTemporaryBan\":{\"Title\":\"Wait for 1 hour\",\"Body\":\"You’ve tried to enter a one-time key too many times.\"},\"OtcUploadDefault\":{\"Title\":\"Try again later\",\"Body\":\"Something went wrong with the system.\"},\"OtcUploadOffline\":{\"Title\":\"Lost connection\",\"Body\":\"Your key was not sent. Make sure your phone is connected to the Internet, through either data or WiFi.\"},\"ContentNoConnect\":{\"Title\":\"Cannot connect to information for your province or territory\",\"Body\":\"Try again later for this information.\"},\"DataSharingStep2\":{\"Title\":\"\",\"Body\":\"You need to choose an answer before going to the next step.\"},\"ApiNotConnected\":{\"Title\":\"Something went wrong\",\"Body\":\"Unfortunately, it looks like COVID Alert will not work on this phone. \\n\\nIf you have more than one profile on your phone, try installing the app in the main profile.\"}}},\"fr\":{\"Home\":{\"AppName\":\"Alerte COVID\",\"UserStopped\":{\"Title\":\"Vous avez désactivé l’application\",\"CTA\":\"Réactiver\"},\"NoExposureDetected\":{\"AllSetTitle\":\"Vous voilà prêt!\",\"RegionCovered\":{\"Title\":\"Aucune exposition détectée\",\"Body\":\"Vous n’avez été près d’aucune personne ayant signalé un diagnostic de COVID-19 avec cette application.\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"AllSetBody\":\"Votre téléphone utilise maintenant Bluetooth pour collecter les codes provenant des téléphones voisins.\"},\"NoRegion\":{\"Title\":\"Aucune exposition détectée\",\"Body\":\"Vous n’avez été près d’aucune personne ayant signalé un diagnostic de COVID-19 avec cette application.\\n\\nToutefois, dans certaines provinces et certains territoires, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\",\"AllSetBody\":\"Nous vous informerons dès que les personnes en dehors de l’Ontario peuvent signaler des diagnostics avec Alerte COVID.\\n\\nIl est tout de même utile de garder Alerte COVID activée.\\n\\nDe cette façon, vous pourrez être avisé dès que les personnes dont vous avez été près signalent un diagnostic.\"},\"RegionNotCovered\":{\"Title\":\"Pas encore de signalement dans votre région\",\"Body\":\"Les personnes de votre province ou territoire ne peuvent pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nIl est tout de même utile d’activer Alerte COVID. Quand les personnes pourront signaler un diagnostic, vous serez ainsi avisé si jamais vous avez été près d’elles.\",\"GuidanceUrl\":\"https://www.canada.ca/fr/sante-publique/services/publications/maladies-et-affections/covid-19-comment-isoler-chez-soi.html\",\"AllSetBody\":\"Nous vous informerons dès que les personnes de votre région peuvent signaler un diagnostic avec Alerte COVID.\\n\\nIl est tout de même utile de garder Alerte COVID activée.\\n\\nDe cette façon, vous pourrez être avisé dès que les personnes dont vous avez été près signalent un diagnostic.\"}},\"ExposureDetected\":{\"Title\":\"Vous avez été exposé\",\"Body1\":\"Vous avez côtoyé au moins une personne qui a signalé un diagnostic de COVID-19 au moyen de l’application. Le contact a duré 15 minutes ou plus.\",\"Title2\":\"Et ensuite?\",\"Body2\":\"Vous pourriez avoir contracté le virus.\",\"Notification\":{\"Received\":\"Date de la notification\"},\"NegativeTest\":{\"CTA\":\"Résultat de test négatif?\"},\"Dismiss\":{\"Title\":\"Résultat négatif au test de dépistage?\",\"Body\":\"Vous pouvez mettre fin à l’alerte pour continuer à recevoir des notifications d’exposition. Aucune information d’exposition ne sera supprimée de votre téléphone.\",\"Body2\":\"Vous pouvez aussi mettre fin à l’alerte si vous suivez les consignes de votre région.\",\"CTA\":\"Mettre fin à l’alerte\",\"Confirm\":{\"Body\":\"Voulez-vous vraiment mettre fin à l’alerte?\",\"Accept\":\"Oui, mettre fin à l’alerte\",\"Cancel\":\"Annuler\"}}},\"ChooseRegionCTA\":\"Sélectionnez une province ou un territoire\",\"How\":\"Diagnosed with COVID-19?\",\"CTA\":\"Prochaines étapes\",\"HowUrl\":\"https://www.canada.ca/\",\"ExposedHelpCTA\":\"Que faire en cas d’exposition\",\"DiagnosedView\":{\"Title\":\"Merci d’aider à lutter contre la COVID\",\"Body1\":{\"One\":\"Demain, vous recevrez une dernière notification vous demandant l’autorisation de téléverser tout nouveau code aléatoire émis par votre téléphone.\",\"Other\":\"Au cours des {number} prochains jours, vous recevrez une notification quotidienne vous demandant l’autorisation de partager toute nouvelle exposition enregistrée par votre téléphone.\"},\"Body2\":\"Vous n’êtes pas tenu d’accepter, même si vous avez accepté le jour d’avant.\",\"Body3\":\"Si vos symptômes vous inquiètent, vous pouvez consulter un professionnel de la santé. Prenez soin de vous et de votre famille pendant ce temps.\",\"Tip\":{\"Title\":\"CONSEIL : \"}},\"SymptomTrackerUrl\":\"https://ca.thrive.health/covid19app/action/88a5e9a7-db9e-487b-bc87-ce35035f8e5c?from=/home&navigateTo=/tracker/complete\",\"DiagnosedShareView\":{\"Title\":\"Partager les nouvelles expositions\",\"Body1\":\"Merci d’aider à ralentir la propagation de la COVID-19!\",\"Body2\":\"Rappel : \",\"Body3\":\"Pour contribuer à la sécurité des autres, vous devez partager vos nouvelles expositions tous les jours, surtout si vous avez à sortir de la maison.\",\"ButtonCTA\":\"Partager les nouvelles expositions\"},\"DiagnosedShareUploadView\":{\"Title\":\"Pas tout à fait fini\",\"Body1\":\"Vous avez entré votre clé à usage unique, mais vous n’avez pas encore partagé vos expositions. Les personnes que vous avez côtoyées n’ont pas encore été notifiées.\",\"ButtonCTA\":\"Finir le partage d’expositions\"},\"BluetoothDisabled\":\"Bluetooth désactivé\",\"EnableBluetoothCTA\":\"Vous devez activer la fonction Bluetooth dans les paramètres de votre téléphone pour qu’Alerte COVID fonctionne.\",\"TurnOnBluetooth\":\"Activez Bluetooth\",\"EnUnauthorized\":{\"Title\":\"Alerte COVID est désactivée\",\"Body1\":\"Appuyez sur le bouton ci-dessous, puis sélectionnez Notifications d’exposition. Faites défiler l’écran et activez « Partager les informations sur les expositions ».\",\"CTA\":\"Activer Alerte COVID\"},\"EnDisabled\":{\"Title\":\"Alerte COVID est désactivée\",\"Body1\":\"Vous n’avez pas activé Alerte COVID. L’application a besoin de votre autorisation pour fonctionner.\",\"CTA\":\"Activer Alerte COVID\",\"AndroidTitle2\":\"Vous hésitez à activer la localisation?\",\"AndroidBody1\":\"Il faut activer la localisation pour qu’Alerte COVID fonctionne. Avec Android, la recherche Bluetooth n’est possible que si la localisation est activée pour toutes les applications.\",\"AndroidBody2a\":\"Malgré cela, Alerte COVID \",\"AndroidBody2b\":\"ne peut pas savoir où vous êtes. \",\"AndroidBody2c\":\"Vous pouvez vous en assurer en vérifiant les autorisations d’Alerte COVID dans les paramètres de votre téléphone. La localisation n’est pas autorisée.\"},\"FrameworkUnavailable\":{\"Title\":\"Une erreur s’est produite\",\"Body\":\"Pour utiliser Alerte COVID, vous devez mettre à jour le système d’exploitation de votre téléphone.\"},\"UnknownProblem\":{\"Title\":\"Une erreur s’est produite\",\"Body\":\"Il semble que votre téléphone ne puisse pas prendre en charge Alerte COVID.\",\"CTA\":\"Trouver de l’aide\"},\"NoConnectivity\":\"Alerte COVID est hors ligne\",\"NoConnectivityDetailed\":\"Alerte COVID vérifie s’il y a des expositions seulement quand votre téléphone est connecté à Internet.\",\"ExternalLinkHint\":\"Ouvre dans une nouvelle fenêtre\",\"LastCheckedMinutes\":{\"One\":\"Dernière vérification il y a {number} minute\",\"Other\":\"Dernière vérification il y a {number} minutes\"},\"LastCheckedHours\":{\"One\":\"Dernière vérification il y a {number} heure\",\"Other\":\"Dernière vérification il y a {number} heures\"},\"LastCheckedDays\":{\"One\":\"Dernière vérification il y a {number} jour\",\"Other\":\"Dernière vérification il y a {number} jours\"}},\"OverlayClosed\":{\"SystemStatus\":\"Alerte COVID est \",\"SystemStatusOn\":\"activée\",\"SystemStatusOff\":\"désactivée\",\"NotificationStatus\":\"Notifications \",\"NotificationStatusOff\":\"désactivées\",\"TapPrompt\":\"Appuyez pour voir le menu\"},\"OverlayOpen\":{\"Version\":\"Version \",\"BluetoothCardBody\":\"Vous devez l’activer dans les paramètres de votre téléphone pour qu’Alerte COVID fonctionne.\",\"BluetoothCardAction\":\"Activez Bluetooth\",\"EnterCodeCardTitle\":\"Avez-vous reçu un diagnostic de COVID-19?\",\"EnterCodeCardBody\":\"Vous devez avoir une clé à usage unique pour aviser les personnes exposées. Nous vous demanderons aussi quelques précisions pour déterminer la période où vous étiez probablement le plus contagieux.\",\"EnterCodeCardAction\":\"Entrer votre clé\",\"EnterCodeCardTitleDiagnosed\":\"Vous contribuez à freiner la COVID\",\"EnterCodeCardBodyDiagnosed\":\"Merci de faire votre part pour freiner la propagation de la COVID-19. \",\"EnterCodeCardDiagnosedCountdown\":{\"One\":\"Il vous reste {number} jour!\",\"Other\":\"Il vous reste {number} jours!\"},\"CodeNotShared\":{\"Title\":\"Aidez à ralentir la propagation de la COVID-19 \",\"Body\":\"Vous avez entré votre clé à usage unique, mais vous n’avez pas encore partagé vos expositions.\",\"CTA\":\"Finir le partage d’expositions\"},\"NotificationCardStatus\":\"Les notifications sont désactivées\",\"NotificationCardStatusOff\":\"DÉSACTIVÉES\",\"NotificationCardBody\":\"Alerte COVID ne peut pas vous aviser dès qu’il découvre une exposition potentielle. Le fait d’être notifié rapidement est essentiel pour ralentir la propagation de la COVID-19.\",\"NotificationCardAction\":\"Activer les notifications\",\"ExposureNotificationCardStatus\":\"Les notifications d’exposition à la COVID-19 sont \",\"ExposureNotificationCardBody\":\"Alerte COVID ne peut pas vous aviser de la découverte d’une exposition potentielle. Le fait d’être notifié rapidement est essentiel pour ralentir la propagation de la COVID-19.\",\"ExposureNotificationCardAction\":\"Activez les notifications d’exposition\",\"NoConnectivityCardAction\":\"Alerte COVID est hors ligne\",\"NoConnectivityCardBody\":\"Connectez-vous à Internet pour aviser les autres personnes ou pour vérifier si vous avez été exposé.\",\"EnUnauthorizedCardAction\":\"Activer Alerte COVID\",\"EnUnauthorizedCardBody\":\"Sélectionnez Notifications d’exposition. Faites défiler l’écran et activez « Partager les informations sur les expositions ».\",\"ShareDiagnosisCode\":{\"CtaReport\":\"Entrer une clé à usage unique\",\"CtaFinish\":\"Finir le partage d’expositions\"}},\"Info\":{\"CheckSymptoms\":\"Vérifier vos symptômes\",\"SymptomsUrl\":\"https://ca.thrive.health/covid19/fr\",\"TellAFriend\":\"Partager l’application\",\"Support\":\"Aide\",\"LearnMore\":\"Fonctionnement\",\"Help\":\"Aide et questions fréquentes\",\"HelpUrl\":\"https://www.canada.ca/fr/sante-publique/services/maladies/maladie-coronavirus-covid-19/alerte-covid/aide.html\",\"GetCode\":\"Obtenir une clé à usage unique\",\"YourVisits\":\"Vos visites\",\"YourRecentExposures\":\"Vos expositions récentes\",\"ChangeLanguage\":\"Modifier la langue\",\"ToggleCovidAlert\":{\"TurnOff\":\"Désactiver Alerte COVID\",\"TurnOn\":\"Réactiver Alerte COVID\",\"Confirm\":{\"Title\":\"Désactiver Alerte COVID?\",\"Body\":\"Veillez à porter un équipement de protection individuelle (ÉPI) médical ou à ne côtoyer que les membres de votre ménage. Une fois désactivée, l’application ne pourra plus enregistrer d’expositions.\",\"Cancel\":\"Ne pas désactiver\",\"Accept\":\"Désactiver\"}},\"ChangeRegion\":\"Modifier la province ou le territoire\",\"Privacy\":\"Confidentialité\",\"PrivacyUrl\":\"https://www.canada.ca/fr/sante-publique/services/maladies/maladie-coronavirus-covid-19/alerte-covid/politique-confidentialite.html\",\"SettingsTitle\":\"Paramètres\",\"InformationTitle\":\"À propos d’Alerte COVID\"},\"Landing\":{\"En\":\"English\",\"Fr\":\"Français\",\"AltText\":\"Illustration de 8 personnes profitant de l’espace public. Il y a une personne avec une canne blanche, une autre en fauteuil roulant, une autre avec un hijab, plusieurs personnes de couleur, et une représentation de la diversité de genre.\",\"CanadaAltText\":\"Symbole du gouvernement du Canada\"},\"LanguageSelect\":{\"Title\":\"Langue\",\"En\":\"English (Canada)\",\"Fr\":\"Français (Canada)\",\"EnShort\":\"English\",\"FrShort\":\"Français\",\"Close\":\"Fermer\"},\"Tutorial\":{\"step-1Title\":\"Fonctionnement d’Alerte COVID\",\"step-1AltText\":\"Dessin à la main d’un téléphone mobile dans une poche, qui envoie et reçoit des signaux.\",\"step-1\":\"L’application s’exécute en arrière-plan sans interrompre vos activités.\\n\\nQuand vous serez près d’une personne ayant Alerte COVID, vos téléphones s’échangeront des codes aléatoires toutes les 5 minutes.\\n\\nCes codes changent souvent et ne peuvent pas servir à vous identifier.\",\"step-2Title\":\"Qu’est-ce qu’une exposition?\",\"step-2AltText\":\"Dessin à la main de 2 téléphones mobiles côte à côte avec un chronomètre qui relie les deux téléphones.\",\"step-2\":\"L’application estime la distance entre deux personnes grâce à la puissance des signaux Bluetooth.\\n\\nPour que l’application enregistre une exposition, un contact doit :\\n• Avoir lieu à moins de 2 mètres.\\n• Durer au moins 15 minutes.\\n• Avoir lieu quand l’autre personne est probablement le plus contagieuse.\",\"step-3Title\":\"En cas de test positif\",\"step-3AltText\":\"Dessin à la main d’un téléphone mobile téléversant un code chiffré vers un nuage.\",\"step-3\":\"Si une personne ayant l’application reçoit un diagnostic de COVID-19, elle peut décider de téléverser les codes aléatoires émis par son téléphone.\\n\\nLa personne peut aussi fournir quelques précisions pour déterminer la période où elle était probablement le plus contagieuse. Cette période indique au téléphone les codes aléatoires qu’il faut téléverser.\\n\\nLes codes aléatoires sont envoyés vers un serveur central. Celui-ci ne reçoit aucun autre renseignement.\",\"step-4Title\":\"Découverte des expositions\",\"step-4AltText\":\"Dessin à la main d’un téléphone mobile émettant une liste de codes chiffrés. L’icône de la loupe de recherche met en évidence l’un des codes de la liste.\",\"step-4\":\"Chaque jour, quand votre téléphone a une connexion Internet, il obtient une liste de codes aléatoires émis par les personnes qui ont signalé un diagnostic.\\n\\nSi l’application reconnaît des codes dans cette liste, elle vous dira que vous avez été exposé et vous expliquera quoi faire ensuite.\",\"ActionBack\":\"Précédent\",\"ActionNext\":\"Suivant\",\"ActionEnd\":\"Terminé\",\"Close\":\"Fermer\"},\"QRCodeOnboarding\":{\"step-1Title\":\"Projet pilote à Mooseville : expositions dans des lieux publics\",\"step-1AltText\":\"-\",\"step-1\":\"L’application peut vous aviser si vous avez visité un lieu pendant qu’il y avait un risque d’exposition.\\n\\nCe type de notifications fonctionne à partir de codes QR que vous scannez lors de vos allées et venues.\\n\\nLes codes QR vous évitent de laisser vos renseignements personnels aux lieux visités, ce qui offre plus de vie privée.\",\"step-2Title\":\"Fonctionnement des notifications par code QR\",\"step-2AltText\":\"-\",\"step-2\":\"Le lieu et le moment où il y a un risque d’exposition est déterminé par les autorités de santé publique.\\n\\nChaque jour, Alerte COVID télécharge une liste de ces lieux et moments.\\n\\nS’il y a correspondance avec l’une de vos visites, l’application vous enverra une notification et vous expliquera quoi faire.\",\"step-3Title\":\"Votre vie privée est protégée\",\"step-3AltText\":\"-\",\"step-3\":\"L’application enregistre la date, l’heure et le lieu de vos visites sur votre téléphone.\\n\\nVos visites et vos expositions **ne sont communiquées à personne**.\\n\\nVous pouvez consulter la liste de vos visites récentes à partir du menu. Au besoin, vous pouvez aussi supprimer des visites.\",\"step-4Title\":\"Facilitons la recherche de contacts\",\"step-4AltText\":\"-\",\"step-4\":\"La prochaine fois que vous sortez, repérez les affiches Alerte COVID avec des codes QR.\\n\\nPuis, commencez à ajouter vos visites et soyez avisé s’il y a eu risque d’exposition pendant celles-ci.\",\"step-4-banner\":\"Pour l’instant, les affiches à code QR ne sont présentes qu’à ________ pour un projet-pilote à petite échelle.\",\"ActionBack\":\"Back\",\"ActionNext\":\"Next\",\"ActionEnd\":\"Done\",\"Close\":\"Close\"},\"Onboarding\":{\"Step\":\"Étape\",\"Of\":\"de\",\"ActionNext\":\"Suivant\",\"ActionEndSkip\":\"Ignorer\",\"ActionEnd\":\"Terminé\",\"ActionBack\":\"Précédent\",\"Start\":{\"Title\":\"Unissons-nous pour freiner la propagation de la COVID-19\",\"ImageAltText\":\"Dessin à la main avec plusieurs personnes différentes dedans. Leurs chemins en lignes pointillées se croisent et mènent à un téléphone intelligent plus bas.\",\"Body1\":\"Alerte COVID nous aide à briser le cycle d’infection. L’application permet d’aviser les personnes en cas d’exposition potentielle avant même que des symptômes apparaissent. \",\"Body2\":\"Nous pouvons ainsi prendre soin de nous et protéger nos communautés.\"},\"WhatItsNot\":{\"Title\":\"Ce qu’Alerte COVID ne fait pas\",\"ImageAltText\":\"Dessin à la main d’une rue sinueuse avec plusieurs maisons et un immeuble en hauteur. De petites infobulles apparaissent au-dessus de chaque bâtiment.\",\"Body1\":\"L’application ne vous dira pas sur-le-champ que vous êtes près d’une personne ayant eu un diagnostic.\",\"Body2\":\"Elle ne vous renseignera pas sur les éclosions de COVID-19 dans votre collectivité.\"},\"Anonymous\":{\"Title\":\"Votre vie privée est protégée\",\"ImageAltText\":\"Dessin à la main d’une paire de lunettes de soleil et d’un chapeau pour passer incognito. Au-dessus du chapeau, une icône de marqueur de géolocalisation est biffée.\",\"Body1\":\"Alerte COVID **n’utilise pas le GPS** et ne suit pas votre emplacement.\",\"Body2\":\"L’application **n’a aucun moyen** de connaître :\",\"Bullet1\":\"Votre emplacement.\",\"Bullet2\":\"Votre nom ou votre adresse.\",\"Bullet3\":\"Les contacts de votre téléphone.\",\"Bullet4\":\"Vos renseignements de santé, sauf ceux que vous acceptez de fournir.\",\"Bullet5\":\"Les renseignements de santé des personnes à proximité.\"},\"HowItWorks\":{\"Title\":\"Fonctionnement\",\"ImageAltText\":\"Dessin à la main de 3 téléphones mobiles émettant des codes aléatoires en utilisant Bluetooth.\",\"Body1\":\"L’application utilise Bluetooth pour échanger des codes aléatoires avec les téléphones à proximité.\",\"Body2\":\"Chaque jour, elle vérifie une liste de codes aléatoires provenant des personnes ayant signalé un test positif à l’application.\",\"Body3\":\"Si vous avez côtoyé l’une de ces personnes au cours des 14 derniers jours, vous recevrez une notification.\",\"HowItWorksCTA\":\"En savoir plus sur le fonctionnement\"},\"PartOf\":{\"Title\":\"L’un des efforts de santé publique\",\"ImageAltText\":\"Dessin à la main de divers moyens de lutter contre la COVID-19, dont un désinfectant pour les mains, un couvre-visage, des mains avec du savon et l’application Alerte COIVID.\",\"Body1\":\"Alerte COVID n’est qu’une partie des efforts de santé publique visant à freiner la propagation de la COVID-19.\",\"Body2\":\"Continuez à suivre les recommandations de santé publique de votre région.\",\"Body3\":\"Alerte COVID ne remplace pas un avis médical. Si vous tombez malade, contactez votre médecin ou un professionnel de la santé.\"},\"Permissions\":{\"Title\":\"L’application demandera votre autorisation\",\"ImageAltText\":\"Dessin à la main d’un téléphone mobile dont l’écran affiche une fenêtre de dialogue avec un X rouge et un crochet vert.\",\"Body1\":\"**Permettez à l’application** de collecter des codes aléatoires quand vous êtes près d’autres téléphones. L’application connaîtra la date, la durée et la puissance de signal associées à ces codes, mais ces données resteront sur votre téléphone. \",\"Body2\":\"Vous devrez aussi **lui permettre de vous envoyer des notifications**.\",\"PrivacyButtonCTA\":\"En savoir plus sur la confidentialité\"}},\"Privacy\":{\"Title\":\"Politique de confidentialité\",\"Close\":\"Fermer\"},\"ThankYou\":{\"Title\":\"Merci d’aider\",\"Body\":\"Les notifications d’exposition sont activées. Si Alerte COVID détecte que vous avez possiblement été exposé(e) à la COVID-19, vous recevrez une notification.\",\"Dismiss\":\"OK\"},\"DataUpload\":{\"Cancel\":\"Annuler\",\"Close\":\"Fermer\",\"Back\":\"Précédent\",\"InfoSmall\":\"Vos identifiants aléatoires ne seront pas partagés, à moins que vous donniez votre permission à l’étape suivante.\",\"ShareToast\":\"Identifiants aléatoires partagés avec succès\",\"StepXofY\":\"Étape {x} de {y}\",\"Today\":\"Aujourd’hui\",\"Yesterday\":\"Hier\",\"Earlier\":\"Il y a plus longtemps\",\"Step0\":{\"Title\":\"Aviser les personnes d’une exposition\",\"List\":{\"1a\":\"1. Vous entrez la clé à usage unique \",\"1b\":\"reçue lors de votre diagnostic.\",\"2a\":\"2. Vous fournissez quelques précisions \",\"2b\":\"qui aideront l’application à déterminer la période où vous étiez probablement le plus contagieux. Vous n’êtes pas obligé de le faire.\",\"3a\":\"3. Vous autorisez votre téléphone \",\"3b\":\"à partager vos codes aléatoires. Ceux-ci permettent d’aviser les personnes que vous avez côtoyées.\"},\"Body1\":\"Les personnes exposées n’obtiendront aucun renseignement sur vous \",\"Body2\":\"ni sur le moment de votre contact. Les personnes recevront uniquement une notification disant qu’elles ont récemment été exposées.\",\"CTA\":\"Suivant\",\"NoCode\":\"Vous n’avez pas de clé à usage unique?\"},\"IntermediateStep\":{\"Title\":\"Finir le partage d’expositions\",\"List\":{\"1a\":\"1. \",\"1b\":\"Fait.\",\"2a\":\"2. Vous fournissez quelques précisions \",\"2b\":\"qui aideront l’application à déterminer la période où vous étiez probablement le plus contagieux. Vous n’êtes pas obligé de le faire.\",\"3a\":\"3. Vous autorisez votre téléphone \",\"3b\":\"à partager vos codes aléatoires. Ceux-ci permettent d’aviser les personnes que vous avez côtoyées.\"},\"Body1\":\"Personne n’obtiendra de renseignements sur vous \",\"Body2\":\"ni sur le moment de votre contact. Les personnes recevront uniquement une notification disant qu’elles ont récemment été exposées.\",\"CTA\":\"Suivant\"},\"Step2\":{\"Title1\":\"Vous sentez-vous malade?\",\"Body1\":\"Symptômes possibles :\",\"Symptoms\":{\"1\":\"Toux ou difficultés respiratoires.\",\"2\":\"Fièvre ou frissons.\",\"3\":\"Fatigue ou faiblesse intense.\",\"4\":\"Courbatures ou mal de tête.\",\"5\":\"Perte soudaine de l’odorat ou du goût.\",\"6\":\"Maux d’estomac.\"},\"Option1\":\"Oui, vous avez eu 1 ou plusieurs symptômes aujourd’hui ou au cours des 14 derniers jours.\",\"Option2\":\"Non, vous n’avez eu aucun symptôme.\",\"Option3\":\"Vous n’êtes pas certain. \",\"Option4\":\"Vous préférez ne pas répondre.\",\"Title2\":\"Pourquoi cette question\",\"Body2\":\"Votre réponse permet de déterminer la période où vous étiez probablement le plus contagieux. L’application sait ainsi quelles expositions il faut partager pour n’aviser que les personnes côtoyées pendant cette période. Votre réponse ne sera jamais échangée ni conservée ailleurs que sur votre téléphone.\",\"CTA\":\"Suivant\"},\"FormView\":{\"Title\":\"Entrez votre clé à usage unique\",\"Body\":\"Entrez la clé reçue lors de votre diagnostic. Afin de prévenir les fausses notifications, vous ne recevrez une clé que si vous avez obtenu un résultat positif au test de dépistage de la COVID-19.\",\"Action\":\"Soumettre la clé\",\"InputLabel\":\"code Alerte COVID\"},\"OtkDiagnosedView\":{\"Title\":\"Vous avez déjà entré votre clé\",\"Body\":\"Il n’est pas nécessaire d’entrer la clé de nouveau.\",\"CTA\":\"Suivant\"},\"ConsentView\":{\"Title\":\"Vous pouvez maintenant téléverser vos codes\",\"Body1\":\"Votre téléphone vous demandera l’autorisation de téléverser vos codes ou « identifiants » aléatoires les plus récents.\",\"Body2a\":\"Personne ne recevra d’information sur vous \",\"Body2b\":\"ou sur le moment de votre proximité.\",\"Body3\":\"Vous pouvez aider à ralentir la propagation de la COVID-19 si vous acceptez.\",\"Action\":\"Accepter\"},\"TekUploadNoDate\":{\"Title\":\"Aviser les personnes côtoyées\",\"Body\":\"Vous n’avez pas fourni de date pour préciser la période où vous étiez probablement le plus contagieux.\\n\\nVous pouvez maintenant partager les expositions enregistrées par l’application pour aviser les personnes côtoyées au cours des 14 derniers jours.\",\"CTA\":\"Partager les expositions\"},\"SymptomOnsetDate\":{\"Title1\":\"Date de vos premiers symptômes\",\"Title2\":\"Pourquoi cette question\",\"Body1\":\"La date des premiers symptômes permet de déterminer la période où vous étiez probablement le plus contagieux. L’application sait ainsi quelles expositions il faut partager pour n’aviser que les personnes côtoyées pendant cette période.\\n\\nCette date ne sera jamais échangée ni conservée ailleurs que sur votre téléphone.\",\"CTA\":\"Partager les expositions\",\"CTA2\":\"Je préfère ne pas répondre\"},\"TestDate\":{\"Title1\":\"Date de votre test\",\"Body1\":\"Entrez la date de votre test de dépistage de la COVID-19.\",\"Title2\":\"Pourquoi cette question\",\"Body2\":\"La date de votre test permet de déterminer la période où vous étiez probablement le plus contagieux. L’application sait ainsi quelles expositions il faut partager pour n’aviser que les personnes côtoyées pendant cette période.\\n\\nCette date ne sera jamais échangée ni conservée ailleurs que sur votre téléphone.\",\"CTA\":\"Partager les expositions\",\"CTA2\":\"Je préfère ne pas répondre\"},\"NoCode\":{\"NoRegion\":{\"Title\":\"Vous n’avez pas choisi de province ou de territoire\",\"Body\":\"Dans certaines provinces et certains territoires, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\\n\\nPour savoir si vous pouvez obtenir une clé à usage unique, vous devez sélectionner votre région.\",\"ChooseRegionCTA\":\"Sélectionner une province ou un territoire\"},\"RegionNotCovered\":{\"Title\":\"Pas encore de signalement dans votre région\",\"Body\":\"Les personnes de votre région ne peuvent pas encore obtenir de clé à usage unique.\\n\\nVous ne pourrez pas aviser les personnes d’une exposition. Mais vous pouvez tout de même vous isoler pendant 14 jours, et ainsi :\",\"Body2\":\"Assurer la sécurité de votre communauté.\",\"Body3\":\"Protéger les personnes et les animaux vivant avec vous.\",\"Body4\":\"Prendre soin de vous.\"}}},\"Notification\":{\"ExposedMessageTitle\":\"Vous avez été exposé\",\"ExposedMessageBody\":\"Vous avez eu un contact étroit avec une personne qui a signalé un diagnostic de COVID-19 au moyen de l’application. Informez-vous sur les étapes à suivre.\",\"OffMessageTitle\":\"Alerte COVID est désactivée\",\"OffMessageBody\":\"Activez Alerte COVID pour être avisé en cas d’exposition potentielle à la COVID-19.\",\"DailyUploadNotificationTitle\":\"Téléversez vos nouveaux codes aléatoires\",\"DailyUploadNotificationBody\":\"Aidez à ralentir la propagation de la COVID-19 en téléversant vos nouveaux codes aléatoires.\",\"ExposureChecksTitle\":\"Alerte COVID\",\"ExposureChecksBody\":\"Vérification d’exposition en cours\",\"ExposureChecksAndroidChannelName\":\"Vérifications d’exposition\",\"PausedMessageTitle\":\"Vous avez désactivé l’application\",\"PausedMessageBody\":\"L’application ne peut pas enregistrer d’expositions puisque vous l’avez désactivée.\",\"OTKNotSharedTitle\":\"Pas tout à fait fini\",\"OTKNotSharedBody\":\"Vous avez entré votre clé à usage unique, mais vous n’avez pas encore partagé vos expositions. Les personnes que vous avez côtoyées n’ont pas encore été notifiées.\",\"AndroidChannelName\":\"Alerte COVID\",\"OutbreakMessageTitle\":\"You have been exposed to an Outbreak\",\"OutbreakMessageBody\":\"There was an outbreak at one of the locations you scanned\"},\"Partners\":{\"Label\":\"Conçu en partenariat avec\"},\"BottomSheet\":{\"Collapse\":\"Fermer\",\"OnStatus\":\"Alerte COVID est ACTIVÉE, Appuyez pour voir le menu\",\"OffStatus\":\"Alerte COVID est DÉSACTIVÉE, Appuyez pour voir le menu\"},\"RegionLanding\":{\"Title\":\"Sélectionnez votre province ou territoire\",\"Body1\":\"Les gouvernements provinciaux et territoriaux s’efforcent de rendre Alerte COVID utilisable partout au Canada.\",\"Body2\":\"Dans certains endroits, on ne peut pas encore signaler de diagnostics de COVID-19 avec cette application.\",\"RegionSelectBtn\":\"Sélectionnez une province ou un territoire\"},\"RegionPicker\":{\"Title\":\"Où habitez-vous? (facultatif)\",\"SettingsTitle\":\"Province ou territoire\",\"Close\":\"Fermer\",\"Body\":\"Sélectionnez votre région pour savoir si vous pouvez signaler un diagnostic à l’aide de l’application.\",\"Skip\":\"Ignorer\",\"GetStarted\":\"Commencer\",\"Optional\":\"* Facultatif\",\"AB\":\"Alberta\",\"BC\":\"Colombie-Britannique\",\"MB\":\"Manitoba\",\"NB\":\"Nouveau-Brunswick\",\"NL\":\"Terre-Neuve-et-Labrador\",\"NT\":\"Territoires du Nord-Ouest\",\"NS\":\"Nouvelle-Écosse\",\"NU\":\"Nunavut\",\"ON\":\"Ontario\",\"PE\":\"Île-du-Prince-Édouard\",\"QC\":\"Québec\",\"SK\":\"Saskatchewan\",\"YT\":\"Yukon\",\"None\":\"Ne rien sélectionner\",\"ExposedHelpLinkNote\":{\"AB\":\"(page en anglais)\",\"BC\":\"(page en anglais)\",\"SK\":\"(page en anglais)\"}},\"RegionPickerNoPT\":{\"Exposed\":{\"Title\":\"Trouvez les consignes de votre région\",\"Body\":\"Les mesures à prendre après une exposition dépendent de votre province ou de votre territoire.\"}},\"A11yList\":{\"Start\":\"Début de liste\",\"End\":\"Fin de liste\",\"Bullet\":\"Puce\"},\"MenuButton\":\"Menu\",\"QRCode\":{\"CTA\":\"Ajouter une visite\",\"Reader\":{\"Title\":\"Placez votre appareil photo devant le code QR\",\"Body\":\"\",\"Back\":\"Précédent\",\"Learn\":\"\"},\"CameraPermissions\":{\"Title\":\"Autoriser l’accès à l’appareil photo\",\"Body\":\"Pour scanner un lieu, l’application doit pouvoir utiliser l’appareil photo.\",\"Title2\":\"Vous hésitez à autoriser l’accès?\",\"Body2\":\"You can also use your phone’s camera app if you prefer not to let the app access your camera.\",\"CTA\":\"Suivant\",\"Cancel\":\"Pas maintenant\"},\"CameraPermissionDenied\":{\"Title\":\"Autorisation refusée\",\"Body\":\"Vous avez refusé qu’Alerte COVID utilise votre appareil photo.\\n\\nPour scanner un lieu, vous devez autoriser Alerte COVID à utiliser l’appareil photo dans les paramètres de votre téléphone.\\n\\nVous pouvez aussi utiliser directement votre appareil photo ou une application de lecture de codes QR.\",\"CTA\":\"Autoriser l’accès dans les paramètres\",\"CTA2\":\"À propos de l’ajout d’une visite\"},\"CameraPermissionAlert\":{\"Title\":\"« Alerte COVID » souhaite accéder à l’appareil photo\",\"Subtitle\":\"Alerte COVID doit accéder à l’appareil photo pour lire le code QR.\",\"Allow\":\"Autoriser\",\"DontAllow\":\"Ne pas autoriser\"},\"LearnAboutQRScan\":{\"Title\":\"Comment ajouter une visite\",\"Body1\":\"1. Autorisez Alerte COVID à accéder à l’appareil photo pour lire le code QR.\",\"Body2\":\"2. Placez votre appareil devant le code.\"},\"Error\":{\"Title\":\"Code QR non reconnu\",\"Body\":\"Vous n’avez peut-être pas scanné un code QR officiel d’Alerte COVID.\\n\\nDemandez à quelqu’un où se trouve l’affiche officielle d’Alerte COVID.\",\"CTA\":\"Réessayer\",\"CTA2\":\"Retour à l’écran d’accueil\"},\"CheckInView\":{\"Title\":\"Merci d’avoir ajouté une visite\",\"Body1\":\"Vous ne serez avisé que si la santé publique détecte un risque d’exposition lié à votre visite. Vos visites \",\"Body2\":\"ne sont communiquées à personne.\\n\\n\",\"Body3\":\"N’oubliez pas de scanner à chaque visite.\",\"Body4\":\"Vous pouvez supprimer cette visite en tout temps en allant dans Vos visites.\",\"CTA1\":\"Retour à l’écran d’accueil\",\"CTA2\":\"Vos visites\",\"MoreInfo\":\"Annuler\"},\"Exposed\":{\"Title\":\"You've been Exposed\"},\"CancelCheckIn\":{\"Title\":\"FR Scan cancelled\"},\"ScanAPlace\":{\"Title\":\"Ajouter une visite: Mooseville seulement\",\"Body\":\"Vous serez avisé si quelqu’un présent pendant votre visite est déclaré positif prochainement.\\n\\nL’application enregistrera vos visites sur votre téléphone, en plus de leur date et heure.\\n\\nVos visites **ne sont communiquées à personne.**\",\"Banner\":\"Pour l’instant, les affiches à code QR ne sont présentes qu’à Mooseville pour un projet-pilote à petite échelle.\",\"CTA\":\"Suivant\",\"CTA2\":\"À propos de l’ajout d’une visite\"},\"ProximityExposure\":\"Passez un test\",\"OutbreakExposed\":{\"Title\":\"Exposition possible lors d’une visite\",\"Body\":\"Une personne déclarée positive était présente lors d’une de vos visites.\",\"CTA\":\"Résultat de test négatif?\",\"SelfIsolate\":{\"Title\":\"Isolez-vous et passez un test\",\"Body\":\"Vous êtes un contact à risque élevé. Si vous avez des symptômes, faites-vous tester immédiatement. Si vous n’en avez pas, attendez 7 jours après la date d’exposition pour vous faire tester.\\n\\nVous devez vous isoler pendant 14 jours à partir de la date d’exposition.\\n\\nPour toute question, contactez le bureau de santé de votre région.\"},\"GetTested\":{\"Title\":\"Passez un test\",\"Body\":\"Vous devez passer un test de dépistage dès que possible.\\n\\nEntre-temps, évitez les contacts avec les personnes vulnérables.\\n\\nPour toute question, contactez le bureau de santé de votre région.\"},\"SelfMonitor\":{\"Title\":\"Évitez les contacts et surveillez vos symptômes\",\"Body\":\"Vous devez éviter les contacts avec les personnes vulnérables et surveiller vos symptômes. Si vous en développez, remplissez une autoévaluation.\\n\\nPour toute question, contactez le bureau de santé de votre région.\"}}},\"PlacesLog\":{\"Title\":\"Vos visites\",\"Body1\":\"Vous serez avisé s’il y a eu un risque d’exposition lors des visites qui suivent.\",\"Body2a\":\"Vos visites ne sont communiquées \",\"Body2b\":\"à personne. Elles sont uniquement conservées sur votre téléphone.\",\"DeleteBtnCTA\":\"Supprimer toutes les visites\",\"DeleteIcon\":\"Supprimer\",\"NoVisit\":\"Aucune visite\",\"Back\":\"Précédent\",\"Alert\":{\"Title\":\"Voulez-vous vraiment supprimer cette visite?\",\"TitleDeleteAll\":\"Voulez-vous vraiment supprimer toutes les visites?\",\"Subtitle\":\"S’il y a un risque d’exposition dans ce lieu, vous n’en serez pas avisé.\",\"SubtitleDeleteAll\":\"S’il y a un risque d’exposition dans un de ces lieux, vous n’en serez pas avisé.\",\"Cancel\":\"Annuler\",\"ConfirmDelete\":\"Oui, supprimer\",\"ConfirmDeleteAll\":\"Oui, tout supprimer\"}},\"ExposureHistory\":{\"Title\":\"Vos expositions récentes\",\"Body\":\"Voici la liste de vos expositions pour les 14 derniers jours.\",\"NoExposures\":\"Aucune exposition\"},\"Errors\":{\"Action\":\"OK\",\"TekUploadNoDate\":{\"Title\":\"\",\"Body\":\"Vous devez sélectionner une date pour pouvoir partager les expositions.\"},\"TekUploadOffline\":{\"Title\":\"Problème de connexion\",\"Body\":\"Vos codes aléatoires n’ont pas été téléversés. Vérifiez que votre téléphone est connecté à Internet au moyen du Wi-Fi ou de vos données.\"},\"TekUploadServer\":{\"Title\":\"Réessayez plus tard\",\"Body\":\"Il y a eu un problème avec le système.\"},\"TekUploadPermission\":{\"Title\":\"Les codes aléatoires n’ont pas pu être téléversés\",\"Body\":\"Vous n’avez pas donné votre autorisation.\"},\"OtcCodeNotEntered\":{\"Title\":\"\",\"Body\":\"Vous devez entrer votre clé pour pouvoir la soumettre.\"},\"OtcUploadInvalidOneTimeCode\":{\"Title\":\"Clé à usage unique non reconnue\",\"Body\":\"Votre clé n’a pas pu être reconnue. Veuillez essayer à nouveau.\"},\"OtcUploadTemporaryBan\":{\"Title\":\"Attendez pendant 1 heure\",\"Body\":\"Vous avez essayé d’entrer une clé à usage unique trop de fois.\"},\"OtcUploadDefault\":{\"Title\":\"Réessayez plus tard\",\"Body\":\"Il y a eu un problème avec le système.\"},\"OtcUploadOffline\":{\"Title\":\"Problème de connexion\",\"Body\":\"Votre clé n’a pas été envoyée. Vérifiez que votre téléphone est connecté à Internet au moyen du Wi-Fi ou de vos données.\"},\"ContentNoConnect\":{\"Title\":\"Impossible d’accéder aux informations de votre province ou de votre territoire\",\"Body\":\"Réessayez d’y accéder plus tard.\"},\"DataSharingStep2\":{\"Title\":\"\",\"Body\":\"Vous devez sélectionner une réponse pour passer à l’étape suivante.\"},\"ApiNotConnected\":{\"Title\":\"Une erreur s'est produite\",\"Body\":\"Malheureusement, il semble qu'Alerte COVID ne fonctionne pas sur votre téléphone. \\n\\nSi vous avez plusieurs profils utilisateur, essayez d'installer l'application dans le profil de base.\"}}}}")
