+++ src/screens/qr/onboarding/TutorialContent.tsx
@@ -6,16 +6,18 @@ import ScrollView from 'rn-faded-scrollview';
 import {Step1} from './views/Step1';
 import {Step2} from './views/Step2';
 import {Step3} from './views/Step3';
+import {Step4} from './views/Step4';
 import {ItemViewProps} from './views/ItemView';
 
-export type TutorialKey = 'step-1' | 'step-2' | 'step-3';
+export type TutorialKey = 'step-1' | 'step-2' | 'step-3' | 'step-4';
 
-export const tutorialData: TutorialKey[] = ['step-1', 'step-2', 'step-3'];
+export const tutorialData: TutorialKey[] = ['step-1', 'step-2', 'step-3', 'step-4'];
 
 const viewComponents: {[key in TutorialKey]: React.ComponentType<Pick<ItemViewProps, 'isActive'>>} = {
   'step-1': Step1,
   'step-2': Step2,
   'step-3': Step3,
+  'step-4': Step4,
 };
 
 export interface TutorialContentProps {
